New code movement to MSVC modules.
Internal changes:
* Move command-line option handling to MSVC/Options.py
* Move vswhere code to MSVC/VSWhere.py
* Move new vs detection code to MSVC/VSDetect.py
* Update test suite accordingly.
