+++ SCons/Tool/MSCommon/__init__.py
@@ -45,9 +45,6 @@ from SCons.Tool.MSCommon.vc import (  # noqa: F401
     msvc_toolset_versions,
     msvc_toolset_versions_spectre,
     msvc_query_version_toolset,
-    vswhere_push_location,
-    msvs_set_channel_default,
-    msvs_get_channel_default,
 )
 
 from SCons.Tool.MSCommon.vs import (  # noqa: F401
@@ -56,6 +53,15 @@ from SCons.Tool.MSCommon.vs import (  # noqa: F401
     merge_default_version,
     msvs_exists,
     query_versions,
+)
+
+from .MSVC.VSWhere import (  # noqa: F401
+    vswhere_push_location,
+)
+
+from .MSVC.VSDetect import (  # noqa: F401
+    msvs_set_channel_default,
+    msvs_get_channel_default,
 )
 
 from .MSVC.Policy import (  # noqa: F401
