+++ SCons/Tool/MSCommon/vc.py
@@ -35,7 +35,6 @@ MS Compilers: Visual C/C++ detection and configuration.
 #   * Assembly
 """
 
-import subprocess
 import os
 import platform
 import sysconfig
@@ -46,25 +45,17 @@ from collections import (
     namedtuple,
     OrderedDict,
 )
-import json
-from functools import cmp_to_key
 
-import SCons.Script
 import SCons.Util
 import SCons.Warnings
-from SCons.Tool import find_program_path
 
 from . import common
 from .common import (
     CONFIG_CACHE,
     debug,
-    debug_extra,
-    DEBUG_ENABLED,
 )
 
 from . import MSVC
-
-from .MSVC.Util import AutoInitialize
 
 from .MSVC.Exceptions import (
     VisualCException,
@@ -92,78 +83,6 @@ class MSVCUseSettingsError(MSVCUserError):
 
 class BatchFileExecutionError(VisualCException):
     pass
-
-# command-line options
-
-class _Options(AutoInitialize):
-
-    debug_extra = None
-
-    options_force = False
-    options_evar = 'SCONS_ENABLE_MSVC_OPTIONS'
-
-    vswhere_option = '--vswhere'
-    vswhere_dest = 'vswhere'
-    vswhere_val = None
-
-    msvs_channel_option = '--msvs-channel'
-    msvs_channel_dest = 'msvs_channel'
-    msvs_channel_val = None
-
-    @classmethod
-    def _setup(cls) -> None:
-
-        enabled = cls.options_force
-        if not enabled:
-            val = os.environ.get(cls.options_evar)
-            enabled = bool(val in MSVC.Config.BOOLEAN_SYMBOLS[True])
-
-        if enabled:
-
-            SCons.Script.AddOption(
-                cls.vswhere_option,
-                nargs=1,
-                dest=cls.vswhere_dest,
-                default=None,
-                type="string",
-                action="store",
-                help='Add vswhere executable located at EXEPATH.',
-                metavar='EXEPATH',
-            )
-
-            cls.vswhere_val = SCons.Script.GetOption(cls.vswhere_dest)
-
-            SCons.Script.AddOption(
-                cls.msvs_channel_option,
-                nargs=1,
-                dest=cls.msvs_channel_dest,
-                default=None,
-                type="string",
-                action="store",
-                help='Set default msvs CHANNEL [Release, Preview, Any].',
-                metavar='CHANNEL',
-            )
-
-            cls.msvs_channel_val = SCons.Script.GetOption(cls.msvs_channel_dest)
-
-        debug(
-            'enabled=%s, vswhere=%s, msvs_channel=%s',
-            enabled, repr(cls.vswhere_val), repr(cls.msvs_channel_val),
-            extra=cls.debug_extra
-        )
-
-    @classmethod
-    def _initialize(cls) -> None:
-        cls.debug_extra = debug_extra(cls)
-        cls._setup()
-
-    @classmethod
-    def vswhere(cls):
-        return cls.vswhere_val, cls.vswhere_option
-
-    @classmethod
-    def msvs_channel(cls):
-        return cls.msvs_channel_val, cls.msvs_channel_option
 
 # undefined object for dict.get() in case key exists and value is None
 UNDEFINED = MSVC.Config.UNDEFINED
@@ -607,48 +526,6 @@ _LE2003_HOST_TARGET_CFG = _host_target_config_factory(
 
 _CL_EXE_NAME = 'cl.exe'
 
-# internal utilities
-
-class _Util(AutoInitialize):
-
-    debug_extra = None
-
-    # cached values
-    _normalized_path = {}
-    _path_exists = {}
-
-    @classmethod
-    def reset(cls) -> None:
-        cls._normalized_path = {}
-        cls._path_exists = {}
-
-    @classmethod
-    def _initialize(cls) -> None:
-        cls.debug_extra = debug_extra(cls)
-        cls.reset()
-
-    # normalized paths
-
-    @classmethod
-    def normalize_path(cls, pval):
-        rval = cls._normalized_path.get(pval, UNDEFINED)
-        if rval == UNDEFINED:
-            rval = MSVC.Util.normalize_path(pval)
-            cls._normalized_path[pval] = rval
-            debug('norm=%s, pval=%s', repr(rval), repr(pval), extra=cls.debug_extra)
-        return rval
-
-    # path existence
-
-    @classmethod
-    def path_exists(cls, pval):
-        rval = cls._path_exists.get(pval, UNDEFINED)
-        if rval == UNDEFINED:
-            rval = os.path.exists(pval)
-            cls._path_exists[pval] = rval
-            debug('exists=%s, pval=%s', rval, repr(pval), extra=cls.debug_extra)
-        return rval
-
 def get_msvc_version_numeric(msvc_version):
     """Get the raw version numbers from a MSVC_VERSION string, so it
     could be cast to float or other numeric values. For example, '14.0Exp'
@@ -852,339 +729,6 @@ _VCVER = [
     "7.0",
     "6.0"]
 
-class _VSConfig:
-
-    # MSVS IDE binaries
-
-    BITFIELD_DEVELOP     = 0b_1000
-    BITFIELD_EXPRESS     = 0b_0100
-    BITFIELD_EXPRESS_WIN = 0b_0010
-    BITFIELD_EXPRESS_WEB = 0b_0001
-
-    EXECUTABLE_MASK = BITFIELD_DEVELOP | BITFIELD_EXPRESS
-
-    VSProgram = namedtuple("VSProgram", [
-        'program',
-        'bitfield'
-    ])
-
-    DEVENV_COM = VSProgram(program='devenv.com', bitfield=BITFIELD_DEVELOP)
-    MSDEV_COM = VSProgram(program='msdev.com', bitfield=BITFIELD_DEVELOP)
-
-    WDEXPRESS_EXE = VSProgram(program='WDExpress.exe', bitfield=BITFIELD_EXPRESS)
-    VCEXPRESS_EXE = VSProgram(program='VCExpress.exe', bitfield=BITFIELD_EXPRESS)
-
-    VSWINEXPRESS_EXE = VSProgram(program='VSWinExpress.exe', bitfield=BITFIELD_EXPRESS_WIN)
-    VWDEXPRESS_EXE = VSProgram(program='VWDExpress.exe', bitfield=BITFIELD_EXPRESS_WEB)
-
-    # MSVS IDE binary
-
-    VSBinaryConfig = namedtuple("VSBinaryConfig", [
-        'pathcomp',  # vs root -> ide dir
-        'programs',  # ide binaries
-    ])
-
-    # MSVS batch file
-
-    VSBatchConfig = namedtuple("VSBatchConfig", [
-        'pathcomp',  # vs root -> batch dir
-        'script',  # vs script
-    ])
-
-    # detect configuration
-
-    DetectConfig = namedtuple("DetectConfig", [
-        'vs_cfg',  # vs configuration
-        'vc_cfg',  # vc configuration
-    ])
-
-    VSDetectConfig = namedtuple("VSDetectConfig", [
-        'root',  # relative path vc dir -> vs root
-        'vs_binary_cfg',  # vs executable
-        'vs_batch_cfg',  # vs batch file
-    ])
-
-    VCDetectConfigRegistry = namedtuple("VCDetectConfigRegistry", [
-        'regkeys',
-    ])
-
-    _VCRegistryKey = namedtuple("_VCRegistryKey", [
-        'hkroot',
-        'key',
-        'is_vsroot',
-        'is_vcforpython',
-    ])
-
-    class VCRegKey(_VCRegistryKey):
-
-        @classmethod
-        def factory(
-            cls, *,
-            key,
-            hkroot=SCons.Util.HKEY_LOCAL_MACHINE,
-            is_vsroot=False,
-            is_vcforpython=False,
-        ):
-
-            regkey = cls(
-                hkroot=hkroot,
-                key=key,
-                is_vsroot=is_vsroot,
-                is_vcforpython=is_vcforpython,
-            )
-
-            return regkey
-
-    _regkey = VCRegKey.factory
-
-    # vs detect configuration: vswhere
-
-    DETECT_VSWHERE = {
-
-        '2022': DetectConfig(  # 14.3
-            vs_cfg=VSDetectConfig(
-                root=os.pardir,
-                vs_binary_cfg=VSBinaryConfig(
-                    pathcomp=r'Common7\IDE',
-                    programs=[DEVENV_COM],
-                ),
-                vs_batch_cfg=VSBatchConfig(
-                    pathcomp=r'Common7\Tools',
-                    script='VsDevCmd.bat',
-                ),
-            ),
-            vc_cfg=None,
-        ),
-
-        '2019': DetectConfig(  # 14.2
-            vs_cfg=VSDetectConfig(
-                root=os.pardir,
-                vs_binary_cfg=VSBinaryConfig(
-                    pathcomp=r'Common7\IDE',
-                    programs=[DEVENV_COM],
-                ),
-                vs_batch_cfg=VSBatchConfig(
-                    pathcomp=r'Common7\Tools',
-                    script='VsDevCmd.bat',
-                ),
-            ),
-            vc_cfg=None,
-        ),
-
-        '2017': DetectConfig(  # 14.1
-            vs_cfg=VSDetectConfig(
-                root=os.pardir,
-                vs_binary_cfg=VSBinaryConfig(
-                    pathcomp=r'Common7\IDE',
-                    programs=[DEVENV_COM, WDEXPRESS_EXE],
-                ),
-                vs_batch_cfg=VSBatchConfig(
-                    pathcomp=r'Common7\Tools',
-                    script='VsDevCmd.bat',
-                ),
-            ),
-            vc_cfg=None,
-        ),
-
-    }
-
-    # vs detect configuration: registry
-
-    DETECT_REGISTRY = {
-
-        '2015': DetectConfig(  # 14.0
-            vs_cfg=VSDetectConfig(
-                root=os.pardir,
-                vs_binary_cfg=VSBinaryConfig(
-                    pathcomp=r'Common7\IDE',
-                    programs=[DEVENV_COM, WDEXPRESS_EXE, VSWINEXPRESS_EXE, VWDEXPRESS_EXE],
-                ),
-                vs_batch_cfg=VSBatchConfig(
-                    pathcomp=r'Common7\Tools',
-                    # VsDevCmd.bat and vsvars32.bat appear to be different
-                    script='vsvars32.bat',
-                ),
-            ),
-            vc_cfg=VCDetectConfigRegistry(
-                regkeys=[
-                    _regkey(key=r'Microsoft\VisualStudio\14.0\Setup\VC\ProductDir'),
-                    _regkey(key=r'Microsoft\WDExpress\14.0\Setup\VS\ProductDir', is_vsroot=True),
-                    _regkey(key=r'Microsoft\VCExpress\14.0\Setup\VC\ProductDir'),  # not populated?
-                ],
-            ),
-        ),
-
-        '2013': DetectConfig(  # 12.0
-            vs_cfg=VSDetectConfig(
-                root=os.pardir,
-                vs_binary_cfg=VSBinaryConfig(
-                    pathcomp=r'Common7\IDE',
-                    programs=[DEVENV_COM, WDEXPRESS_EXE, VSWINEXPRESS_EXE, VWDEXPRESS_EXE],
-                ),
-                vs_batch_cfg=VSBatchConfig(
-                    pathcomp=r'Common7\Tools',
-                    # VsDevCmd.bat and vsvars32.bat appear to be different
-                    script='vsvars32.bat',
-                ),
-            ),
-            vc_cfg=VCDetectConfigRegistry(
-                regkeys=[
-                    _regkey(key=r'Microsoft\VisualStudio\12.0\Setup\VC\ProductDir'),
-                    _regkey(key=r'Microsoft\VCExpress\12.0\Setup\VC\ProductDir'),
-                ]
-            ),
-        ),
-
-        '2012': DetectConfig(  # 11.0
-            vs_cfg=VSDetectConfig(
-                root=os.pardir,
-                vs_binary_cfg=VSBinaryConfig(
-                    pathcomp=r'Common7\IDE',
-                    programs=[DEVENV_COM, WDEXPRESS_EXE, VSWINEXPRESS_EXE, VWDEXPRESS_EXE],
-                ),
-                vs_batch_cfg=VSBatchConfig(
-                    pathcomp=r'Common7\Tools',
-                    # VsDevCmd.bat and vsvars32.bat appear to be identical
-                    script='vsvars32.bat',
-                ),
-            ),
-            vc_cfg=VCDetectConfigRegistry(
-                regkeys=[
-                    _regkey(key=r'Microsoft\VisualStudio\11.0\Setup\VC\ProductDir'),
-                    _regkey(key=r'Microsoft\VCExpress\11.0\Setup\VC\ProductDir'),
-                ]
-            ),
-        ),
-
-        '2010': DetectConfig(  # 10.0
-            vs_cfg=VSDetectConfig(
-                root=os.pardir,
-                vs_binary_cfg=VSBinaryConfig(
-                    pathcomp=r'Common7\IDE',
-                    programs=[DEVENV_COM, VCEXPRESS_EXE],
-                ),
-                vs_batch_cfg=VSBatchConfig(
-                    pathcomp=r'Common7\Tools',
-                    script='vsvars32.bat',
-                ),
-            ),
-            vc_cfg=VCDetectConfigRegistry(
-                regkeys=[
-                    _regkey(key=r'Microsoft\VisualStudio\10.0\Setup\VC\ProductDir'),
-                    _regkey(key=r'Microsoft\VCExpress\10.0\Setup\VC\ProductDir'),
-                ]
-            ),
-        ),
-
-        '2008': DetectConfig(  # 9.0
-            vs_cfg=VSDetectConfig(
-                root=os.pardir,
-                vs_binary_cfg=VSBinaryConfig(
-                    pathcomp=r'Common7\IDE',
-                    programs=[DEVENV_COM, VCEXPRESS_EXE],
-                ),
-                vs_batch_cfg=VSBatchConfig(
-                    pathcomp=r'Common7\Tools',
-                    script='vsvars32.bat',
-                ),
-            ),
-            vc_cfg=VCDetectConfigRegistry(
-                regkeys=[
-                    _regkey(
-                        hkroot=SCons.Util.HKEY_CURRENT_USER,
-                        key=r'Microsoft\DevDiv\VCForPython\9.0\InstallDir',
-                        is_vsroot=True, is_vcforpython=True,
-                    ),
-                    _regkey(
-                        key=r'Microsoft\DevDiv\VCForPython\9.0\InstallDir',
-                        is_vsroot=True, is_vcforpython=True
-                    ),
-                    _regkey(key=r'Microsoft\VisualStudio\9.0\Setup\VC\ProductDir'),
-                    _regkey(key=r'Microsoft\VCExpress\9.0\Setup\VC\ProductDir'),
-                ]
-            ),
-        ),
-
-        '2005': DetectConfig(  # 8.0
-            vs_cfg=VSDetectConfig(
-                root=os.pardir,
-                vs_binary_cfg=VSBinaryConfig(
-                    pathcomp=r'Common7\IDE',
-                    programs=[DEVENV_COM, VCEXPRESS_EXE],
-                ),
-                vs_batch_cfg=VSBatchConfig(
-                    pathcomp=r'Common7\Tools',
-                    script='vsvars32.bat',
-                ),
-            ),
-            vc_cfg=VCDetectConfigRegistry(
-                regkeys=[
-                    _regkey(key=r'Microsoft\VisualStudio\8.0\Setup\VC\ProductDir'),
-                    _regkey(key=r'Microsoft\VCExpress\8.0\Setup\VC\ProductDir'),
-                ]
-            ),
-        ),
-
-        '2003': DetectConfig(  # 7.1
-            vs_cfg=VSDetectConfig(
-                root=os.pardir,
-                vs_binary_cfg=VSBinaryConfig(
-                    pathcomp=r'Common7\IDE',
-                    programs=[DEVENV_COM],
-                ),
-                vs_batch_cfg=VSBatchConfig(
-                    pathcomp=r'Common7\Tools',
-                    script='vsvars32.bat',
-                ),
-            ),
-                vc_cfg=VCDetectConfigRegistry(
-                regkeys=[
-                    _regkey(key=r'Microsoft\VisualStudio\7.1\Setup\VC\ProductDir'),
-                ]
-            ),
-        ),
-
-        '2002': DetectConfig(  # 7.0
-            vs_cfg=VSDetectConfig(
-                root=os.pardir,
-                vs_binary_cfg=VSBinaryConfig(
-                    pathcomp=r'Common7\IDE',
-                    programs=[DEVENV_COM],
-                ),
-                vs_batch_cfg=VSBatchConfig(
-                    pathcomp=r'Common7\Tools',
-                    script='vsvars32.bat',
-                ),
-            ),
-            vc_cfg=VCDetectConfigRegistry(
-                regkeys=[
-                    _regkey(key=r'Microsoft\VisualStudio\7.0\Setup\VC\ProductDir'),
-                ]
-            ),
-        ),
-
-        '1998': DetectConfig(  # 6.0
-            vs_cfg=VSDetectConfig(
-                root=os.pardir,
-                vs_binary_cfg=VSBinaryConfig(
-                    pathcomp=r'Common\MSDev98\Bin',
-                    programs=[MSDEV_COM],
-                ),
-                vs_batch_cfg=VSBatchConfig(
-                    # There is no vsvars32.bat batch file
-                    pathcomp=r'VC98\bin',
-                    script='vcvars32.bat',
-                ),
-            ),
-            vc_cfg=VCDetectConfigRegistry(
-                regkeys=[
-                    _regkey(key=r'Microsoft\VisualStudio\6.0\Setup\Microsoft Visual C++\ProductDir'),
-                ]
-            ),
-        ),
-
-    }
-
 def msvc_version_to_maj_min(msvc_version):
     msvc_version_numeric = get_msvc_version_numeric(msvc_version)
 
@@ -1198,28 +742,12 @@ def msvc_version_to_maj_min(msvc_version):
     except ValueError:
         raise ValueError("Unrecognized version %s (%s)" % (msvc_version,msvc_version_numeric)) from None
 
-VSWHERE_PATHS = [
-    os.path.join(p,'vswhere.exe')
-    for p in [
-        # For bug 3333: support default location of vswhere for both
-        # 64 and 32 bit windows installs.
-        # For bug 3542: also accommodate not being on C: drive.
-        os.path.expandvars(r"%ProgramFiles(x86)%\Microsoft Visual Studio\Installer"),
-        os.path.expandvars(r"%ProgramFiles%\Microsoft Visual Studio\Installer"),
-        os.path.expandvars(r"%ChocolateyInstall%\bin"),
-        os.path.expandvars(r"%LOCALAPPDATA%\Microsoft\WinGet\Links"),
-        os.path.expanduser(r"~\scoop\shims"),
-        os.path.expandvars(r"%SCOOP%\shims"),
-    ]
-    if not p.startswith('%')
-]
-
 def msvc_find_vswhere(env=None):
     """ Find the location of vswhere """
     # NB: this gets called from testsuite on non-Windows platforms.
     # Whether that makes sense or not, don't break it for those.
     vswhere_env = env.subst('$VSWHERE') if env and 'VSWHERE' in env else None
-    vswhere_executables = _VSWhere.find_executables(vswhere_env)
+    vswhere_executables = MSVC.VSWhere.vswhere_get_executables(vswhere_env)
     if vswhere_executables:
         vswhere_path = vswhere_executables[0].path
     else:
@@ -1227,2176 +755,10 @@ def msvc_find_vswhere(env=None):
     debug('vswhere_path=%s', vswhere_path)
     return vswhere_path
 
-class _VSWhere(AutoInitialize):
-
-    # TODO(JCB): review reset
-
-    # priority: env > cmdline > (user, initial, user)
-
-    VSWHERE_EXE = 'vswhere.exe'
-
-    VSWhereExecutable = namedtuple('VSWhereExecutable', [
-        'path',
-        'norm',
-    ])
-
-    debug_extra = None
-
-    vswhere_cmdline = None
-    vswhere_executables = []
-
-    @classmethod
-    def _cmdline(cls) -> None:
-        vswhere, option = _Options.vswhere()
-        if vswhere:
-            vswhere_exec = cls._user_path(vswhere, option)
-            if vswhere_exec:
-                cls.vswhere_cmdline = vswhere_exec
-        debug('vswhere_cmdline=%s', cls.vswhere_cmdline, extra=cls.debug_extra)
-
-    @classmethod
-    def _setup(cls) -> None:
-        for pval in VSWHERE_PATHS:
-            if not _Util.path_exists(pval):
-                continue
-            vswhere_exec = cls.VSWhereExecutable(path=pval, norm=_Util.normalize_path(pval))
-            cls.vswhere_executables.append(vswhere_exec)
-        debug('vswhere_executables=%s', cls.vswhere_executables, extra=cls.debug_extra)
-        cls._cmdline()
-
-    @classmethod
-    def _initialize(cls) -> None:
-        cls.debug_extra = debug_extra(cls)
-        cls.reset()
-        cls._setup()
-
-    @classmethod
-    def reset(cls) -> None:
-        cls._cache_user_vswhere_paths = {}
-
-    # validate user-specified vswhere executable
-
-    @classmethod
-    def _user_path(cls, pval, source):
-
-        rval = cls._cache_user_vswhere_paths.get(pval, UNDEFINED)
-        if rval != UNDEFINED:
-            debug('vswhere_exec=%s', repr(rval), extra=cls.debug_extra)
-            return rval
-
-        vswhere_exec = None
-        if pval:
-
-            if not _Util.path_exists(pval):
-
-                warn_msg = f'vswhere executable path not found: {pval!r} ({source})'
-                debug(warn_msg, extra=cls.debug_extra)
-                SCons.Warnings.warn(MSVC.Warnings.VSWherePathWarning, warn_msg)
-
-            else:
-
-                norm = _Util.normalize_path(pval)
-                tail = os.path.split(norm)[-1]
-                if tail != cls.VSWHERE_EXE:
-
-                    warn_msg = f'unsupported vswhere executable (expected {cls.VSWHERE_EXE!r}, found {tail!r}): {pval!r} ({source})'
-                    debug(warn_msg, extra=cls.debug_extra)
-                    SCons.Warnings.warn(MSVC.Warnings.VSWherePathWarning, warn_msg)
-
-                else:
-
-                    vswhere_exec = cls.VSWhereExecutable(path=pval, norm=norm)
-                    debug('vswhere_exec=%s', repr(vswhere_exec), extra=cls.debug_extra)
-
-        cls._cache_user_vswhere_paths[pval] = vswhere_exec
-
-        return vswhere_exec
-
-    # user-specified vswhere executables
-
-    @classmethod
-    def user_pathlist(cls, path_list, front, source) -> None:
-
-        user_executables = []
-        for pval in path_list:
-            vswhere_exec = cls._user_path(pval, source)
-            if vswhere_exec:
-                user_executables.append(vswhere_exec)
-
-        if user_executables:
-
-            if front:
-                all_executables = user_executables + cls.vswhere_executables
-            else:
-                all_executables = cls.vswhere_executables + user_executables
-
-            seen = set()
-            unique_executables = []
-            for vswhere_exec in all_executables:
-                if vswhere_exec.norm in seen:
-                    continue
-                seen.add(vswhere_exec.norm)
-                unique_executables.append(vswhere_exec)
-
-            cls.vswhere_executables = unique_executables
-            debug('vswhere_executables=%s', cls.vswhere_executables, extra=cls.debug_extra)
-
-    # all vswhere executables
-
-    @classmethod
-    def find_executables(cls, vswhere_env=None):
-
-        vswhere_executables = []
-
-        # env['VSWHERE'] path
-        if vswhere_env:
-            vswhere_exec = cls._user_path(vswhere_env, "env['VSWHERE']")
-            if vswhere_exec:
-                vswhere_executables.append(vswhere_exec)
-
-        # --vswhere=EXEPATH
-        if cls.vswhere_cmdline:
-            vswhere_executables.append(cls.vswhere_cmdline)
-
-        # default paths and user paths (vswhere_push_location)
-        if cls.vswhere_executables:
-            vswhere_executables.extend(cls.vswhere_executables)
-
-        return vswhere_executables
-
-# register user vswhere executable location(s)
-
-def vswhere_push_location(string_or_list, front=False) -> None:
-    # TODO(JCB): need docstring
-    path_list = SCons.Util.flatten(string_or_list)
-    if path_list:
-        _VSWhere.user_pathlist(path_list, front, 'vswhere_push_location')
-
-class _VSChannel(AutoInitialize):
-
-    # TODO(JCB): review reset
-
-    # priority: cmdline > user > initial
-
-    debug_extra = None
-
-    vs_channel_initial = None
-    vs_channel_cmdline = None
-
-    # reset
-
-    vs_channel_user = None
-    vs_channel_def = None
-    vs_channel_retrieved = False
-
-    @classmethod
-    def _initial(cls) -> None:
-        cls.vs_channel_initial = MSVC.Config.MSVS_CHANNEL_RELEASE
-        cls.vs_channel_def = cls.vs_channel_initial
-        debug('vs_channel_initial=%s',
-            cls.vs_channel_initial.vs_channel_id,
-            extra=cls.debug_extra
-        )
-
-    @classmethod
-    def _cmdline(cls) -> None:
-        channel, option = _Options.msvs_channel()
-        if channel:
-            vs_channel_def = MSVC.Validate.validate_msvs_channel(
-                channel, option
-            )
-            if vs_channel_def:
-                cls.vs_channel_cmdline = vs_channel_def
-                cls.vs_channel_def = cls.vs_channel_cmdline
-                debug('vs_channel_cmdline=%s',
-                    cls.vs_channel_cmdline.vs_channel_id,
-                    extra=cls.debug_extra
-                )
-
-    @classmethod
-    def _setup(cls) -> None:
-        cls._initial()
-        cls._cmdline()
-
-    @classmethod
-    def _initialize(cls) -> None:
-        cls.debug_extra = debug_extra(cls)
-        cls._setup()
-
-    @classmethod
-    def reset(cls) -> None:
-        cls.vs_channel_user = None
-        cls.vs_channel_def = None
-        cls.vs_channel_retrieved = False
-        cls._setup()
-
-    @classmethod
-    def get_default_channel(cls):
-        if not cls.vs_channel_retrieved:
-            cls.vs_channel_retrieved = True
-            if cls.vs_channel_cmdline:
-                cls.vs_channel_def = cls.vs_channel_cmdline
-            elif cls.vs_channel_user:
-                cls.vs_channel_def = cls.vs_channel_user
-            else:
-                cls.vs_channel_def = cls.vs_channel_initial
-        debug(
-            'vs_channel=%s',
-            cls.vs_channel_def.vs_channel_id,
-            extra=cls.debug_extra
-        )
-        return cls.vs_channel_def
-
-    @classmethod
-    def set_default_channel(cls, msvs_channel, source) -> bool:
-
-        rval = False
-
-        if cls.vs_channel_retrieved:
-
-            warn_msg = f'msvs channel {msvs_channel!r} ({source}) ignored: must be set before first use.'
-            debug(warn_msg, extra=cls.debug_extra)
-            SCons.Warnings.warn(MSVC.Warnings.MSVSChannelWarning, warn_msg)
-
-        else:
-
-            vs_channel_def = MSVC.Validate.validate_msvs_channel(msvs_channel, source)
-            if vs_channel_def:
-
-                cls.vs_channel_user = vs_channel_def
-                debug(
-                    'vs_channel_user=%s',
-                    cls.vs_channel_user.vs_channel_id,
-                    extra=cls.debug_extra
-                )
-
-                if not cls.vs_channel_cmdline:
-                    # priority: cmdline > user > initial
-                    cls.vs_channel_def = cls.vs_channel_user
-                    debug(
-                        'vs_channel=%s',
-                        cls.vs_channel_def.vs_channel_id,
-                        extra=cls.debug_extra
-                    )
-                    rval = True
-
-                debug(
-                    'vs_channel=%s',
-                    cls.vs_channel_def.vs_channel_id,
-                    extra=cls.debug_extra
-                )
-
-        return rval
-
-def msvs_set_channel_default(msvs_channel) -> bool:
-    """Set the default msvs channel.
-
-    Args:
-        msvs_channel: str
-            string representing the msvs channel
-
-            Case-insensitive values are:
-                Release, Rel, Preview, Pre, Any, *
-
-    Returns:
-        bool: True if the default msvs channel was accepted.
-              False if the default msvs channel was not accepted.
-
-    """
-    rval = _VSChannel.set_default_channel(msvs_channel, 'msvc_set_channel_default')
-    return rval
-
-def msvs_get_channel_default():
-    """Get the default msvs channel.
-
-    Returns:
-        str: default msvs channel
-
-    """
-    return _VSChannel.vs_channel_def.vs_channel_id
-
-class _VSKeys(AutoInitialize):
-
-    debug_extra = None
-
-    @classmethod
-    def _initialize(cls) -> None:
-        cls.debug_extra = debug_extra(cls)
-
-    # edition key: (product, channel, component/None, seqnbr/None)
-
-    _MSVSEditionKey = namedtuple('_MSVSEditionKey', [
-        'vs_product_def',
-        'vs_channel_def',
-        'vs_componentid_def',
-        'vs_sequence_nbr',
-    ])
-
-    class MSVSEditionKey(_MSVSEditionKey):
-
-        def serialize(self):
-            values = [
-                self.vs_product_def.vs_product,
-                self.vs_channel_def.vs_channel_suffix,
-            ]
-            if self.vs_componentid_def:
-                values.append(self.vs_componentid_def.vs_component_suffix)
-            if self.vs_sequence_nbr:
-                values.append(str(self.vs_sequence_nbr))
-            rval = '-'.join(values)
-            return rval
-
-        @classmethod
-        def factory(
-            cls, *,
-            vs_product_def,
-            vs_channel_def,
-            vs_componentid_def,
-            vs_sequence_nbr,
-        ):
-
-            vs_edition_key = cls(
-                vs_product_def=vs_product_def,
-                vs_channel_def=vs_channel_def,
-                vs_componentid_def=vs_componentid_def,
-                vs_sequence_nbr=vs_sequence_nbr
-            )
-
-            return vs_edition_key
-
-    @classmethod
-    def msvs_edition_key(
-        cls, *,
-        vs_product_def=None,
-        vs_channel_def=None,
-        vs_componentid_def=None,
-        vs_sequence_nbr=None,
-    ):
-
-        if not vs_product_def:
-            errmsg = 'vs_product_def is undefined'
-            debug('MSVCInternalError: %s', errmsg, extra=cls.debug_extra)
-            raise MSVCInternalError(errmsg)
-
-        if not vs_channel_def:
-            errmsg = 'vs_channel_def is undefined'
-            debug('MSVCInternalError: %s', errmsg, extra=cls.debug_extra)
-            raise MSVCInternalError(errmsg)
-
-        vs_edition_key = cls.MSVSEditionKey.factory(
-            vs_product_def=vs_product_def,
-            vs_channel_def=vs_channel_def,
-            vs_componentid_def=vs_componentid_def,
-            vs_sequence_nbr=vs_sequence_nbr
-        )
-
-        return vs_edition_key
-
-    # channel key: (channel, component/None)
-
-    _MSVSChannelKey = namedtuple('_MSVSChannelKey', [
-        'vs_channel_def',
-        'vs_componentid_def',
-    ])
-
-    class MSVSChannelKey(_MSVSChannelKey):
-
-        def serialize(self):
-            values = [
-                self.vs_channel_def.vs_channel_suffix,
-            ]
-            if self.vs_componentid_def:
-                values.append(self.vs_componentid_def.vs_component_suffix)
-            rval = '-'.join(values)
-            return rval
-
-        @classmethod
-        def factory(
-            cls, *,
-            vs_channel_def,
-            vs_componentid_def,
-        ):
-
-            vs_channel_key = cls(
-                vs_channel_def=vs_channel_def,
-                vs_componentid_def=vs_componentid_def,
-            )
-
-            return vs_channel_key
-
-    @classmethod
-    def msvs_channel_key(
-        cls, *,
-        vs_channel_def=None,
-        vs_componentid_def=None,
-    ):
-
-        if not vs_channel_def:
-            errmsg = 'vs_channel_def is undefined'
-            debug('MSVCInternalError: %s', errmsg, extra=cls.debug_extra)
-            raise MSVCInternalError(errmsg)
-
-        vs_channel_key = cls.MSVSChannelKey.factory(
-            vs_channel_def=vs_channel_def,
-            vs_componentid_def=vs_componentid_def,
-        )
-
-        return vs_channel_key
-
-_MSVSBase = namedtuple('_MSVSBase', [
-    'id_str',
-    'id_comps',
-    'vs_product_def',
-    'vs_channel_def',
-    'vs_component_def',
-    'vs_sequence_nbr',
-    'vs_dir',
-    'vs_dir_norm',
-    'vs_version',
-    'is_express',
-    'is_buildtools',
-    'is_vcforpython',
-    'vs_edition_channel_component_seqnbr_key',
-    'vs_edition_channel_component_key',
-    'vs_edition_channel_key',
-    'vs_channel_component_key',
-    'vs_channel_key',
-    'instance_map',
-])
-
-class MSVSBase(_MSVSBase):
-
-    def _is_express(vs_component_def) -> bool:
-        vs_componentid_def = vs_component_def.vs_componentid_def
-        is_express = bool(vs_componentid_def == MSVC.Config.MSVS_COMPONENTID_EXPRESS)
-        return is_express
-
-    @staticmethod
-    def _is_buildtools(vs_component_def) -> bool:
-        vs_componentid_def = vs_component_def.vs_componentid_def
-        is_buildtools = bool(vs_componentid_def == MSVC.Config.MSVS_COMPONENTID_BUILDTOOLS)
-        return is_buildtools
-
-    @staticmethod
-    def _is_vcforpython(vs_component_def) -> bool:
-        vs_componentid_def = vs_component_def.vs_componentid_def
-        is_vcforpython = bool(vs_componentid_def == MSVC.Config.MSVS_COMPONENTID_PYTHON)
-        return is_vcforpython
-
-    @classmethod
-    def factory(
-        cls, *,
-        vs_product_def,
-        vs_channel_def,
-        vs_component_def,
-        vs_sequence_nbr,
-        vs_dir,
-        vs_version,
-    ):
-
-        vs_componentid_def = vs_component_def.vs_componentid_def
-
-        vs_edition_channel_component_seqnbr_key = _VSKeys.msvs_edition_key(
-            vs_product_def=vs_product_def,
-            vs_channel_def=vs_channel_def,
-            vs_componentid_def=vs_componentid_def,
-            vs_sequence_nbr=vs_sequence_nbr,
-        )
-
-        vs_edition_channel_component_key = _VSKeys.msvs_edition_key(
-            vs_product_def=vs_product_def,
-            vs_channel_def=vs_channel_def,
-            vs_componentid_def=vs_componentid_def,
-        )
-
-        vs_edition_channel_key = _VSKeys.msvs_edition_key(
-            vs_product_def=vs_product_def,
-            vs_channel_def=vs_channel_def,
-        )
-
-        vs_channel_component_key = _VSKeys.msvs_channel_key(
-            vs_channel_def=vs_channel_def,
-            vs_componentid_def=vs_componentid_def,
-        )
-
-        vs_channel_key = _VSKeys.msvs_channel_key(
-            vs_channel_def=vs_channel_def,
-        )
-
-        instance_map = {
-            'msvs_instance': None,
-            'msvc_instance': None,
-        }
-
-        id_comps = (
-            vs_product_def.vs_product,
-            vs_channel_def.vs_channel_suffix,
-            vs_component_def.vs_componentid_def.vs_component_suffix,
-            str(vs_sequence_nbr),
-        )
-
-        id_str = '{}({})'.format(cls.__name__, ', '.join(id_comps))
-
-        msvs_base = cls(
-            id_str=id_str,
-            id_comps=id_comps,
-            vs_product_def=vs_product_def,
-            vs_channel_def=vs_channel_def,
-            vs_component_def=vs_component_def,
-            vs_sequence_nbr=vs_sequence_nbr,
-            vs_dir=vs_dir,
-            vs_dir_norm=_Util.normalize_path(vs_dir),
-            vs_version=vs_version,
-            is_express=cls._is_express(vs_component_def),
-            is_buildtools=cls._is_buildtools(vs_component_def),
-            is_vcforpython=cls._is_vcforpython(vs_component_def),
-            vs_edition_channel_component_seqnbr_key=vs_edition_channel_component_seqnbr_key,
-            vs_edition_channel_component_key=vs_edition_channel_component_key,
-            vs_edition_channel_key=vs_edition_channel_key,
-            vs_channel_component_key=vs_channel_component_key,
-            vs_channel_key=vs_channel_key,
-            instance_map=instance_map,
-        )
-
-        return msvs_base
-
-    @staticmethod
-    def default_order(a, b):
-        # vs product numeric: descending order
-        if a.vs_product_def.vs_product_numeric != b.vs_product_def.vs_product_numeric:
-            return 1 if a.vs_product_def.vs_product_numeric < b.vs_product_def.vs_product_numeric else -1
-        # vs channel: ascending order (release, preview)
-        if a.vs_channel_def.vs_channel_rank != b.vs_channel_def.vs_channel_rank:
-            return 1 if a.vs_channel_def.vs_channel_rank > b.vs_channel_def.vs_channel_rank else -1
-        # component rank: descending order
-        if a.vs_component_def.vs_component_rank != b.vs_component_def.vs_component_rank:
-            return 1 if a.vs_component_def.vs_component_rank < b.vs_component_def.vs_component_rank else -1
-        # sequence number: ascending order
-        if a.vs_sequence_nbr != b.vs_sequence_nbr:
-            return 1 if a.vs_sequence_nbr > b.vs_sequence_nbr else -1
-        return 0
-
-    def register_msvs_instance(self, msvs_instance) -> None:
-        self.instance_map['msvs_instance'] = msvs_instance
-
-    def register_msvc_instance(self, msvc_instance) -> None:
-        self.instance_map['msvc_instance'] = msvc_instance
-
-    @property
-    def msvs_instance(self):
-        return self.instance_map.get('msvs_instance')
-
-    @property
-    def msvc_instance(self):
-        return self.instance_map.get('msvc_instance')
-
-    @property
-    def vs_component_suffix(self):
-        return self.vs_component_def.vs_componentid_def.vs_component_suffix
-
-_MSVSInstance = namedtuple('_MSVSInstance', [
-    'id_str',
-    'msvs_base',
-    'vs_executable',
-    'vs_executable_norm',
-    'vs_script',
-    'vs_script_norm',
-    'vc_version_def',
-])
-
-class MSVSInstance(_MSVSInstance):
-
-    @classmethod
-    def factory(
-        cls, *,
-        msvs_base,
-        vs_executable,
-        vs_script,
-        vc_version_def,
-    ):
-
-        id_str = '{}({})'.format(cls.__name__, ', '.join(msvs_base.id_comps))
-
-        msvs_instance = cls(
-            id_str=id_str,
-            msvs_base=msvs_base,
-            vs_executable=vs_executable,
-            vs_executable_norm=_Util.normalize_path(vs_executable),
-            vs_script=vs_script,
-            vs_script_norm=_Util.normalize_path(vs_script),
-            vc_version_def=vc_version_def,
-        )
-
-        return msvs_instance
-
-    @staticmethod
-    def default_order(a, b):
-        return MSVSBase.default_order(a.msvs_base, b.msvs_base)
-
-    @property
-    def id_comps(self):
-        return self.msvs_base.id_comps
-
-    @property
-    def msvc_instance(self):
-        return self.msvs_base.msvc_instance
-
-    @property
-    def vs_dir(self):
-        return self.msvs_base.vs_dir
-
-    @property
-    def vs_version(self):
-        return self.msvs_base.vs_version
-
-    @property
-    def vs_edition_channel_component_seqnbr_key(self):
-        return self.msvs_base.vs_edition_channel_component_seqnbr_key
-
-    @property
-    def vs_edition_channel_component_key(self):
-        return self.msvs_base.vs_edition_channel_component_key
-
-    @property
-    def vs_edition_channel_key(self):
-        return self.msvs_base.vs_edition_channel_key
-
-    @property
-    def vs_channel_component_key(self):
-        return self.msvs_base.vs_channel_component_key
-
-    @property
-    def vs_channel_key(self):
-        return self.msvs_base.vs_channel_key
-
-    @property
-    def msvc_version(self):
-        return self.vc_version_def.msvc_version
-
-    @property
-    def msvc_verstr(self):
-        return self.vc_version_def.msvc_verstr
-
-    @property
-    def vs_product_def(self):
-        return self.msvs_base.vs_product_def
-
-    @property
-    def vs_channel_def(self):
-        return self.msvs_base.vs_channel_def
-
-    @property
-    def vs_componentid_def(self):
-        return self.msvs_base.vs_component_def.vs_componentid_def
-
-    @property
-    def vs_product(self):
-        return self.msvs_base.vs_product_def.vs_product
-
-    @property
-    def vs_channel_id(self):
-        return self.msvs_base.vs_channel_def.vs_channel_id
-
-    @property
-    def vs_component_id(self):
-        return self.msvs_base.vs_component_def.vs_componentid_def.vs_component_id
-
-    @property
-    def vs_sequence_nbr(self):
-        return self.msvs_base.vs_sequence_nbr
-
-_MSVSInstalled = namedtuple('_MSVSInstalled', [
-    'msvs_instances',
-    'msvs_edition_instances_map',
-    'msvs_channel_instances_map',
-    'msvs_channel_map', # TODO(JCB): remove?
-])
-
-class MSVSInstalled(_MSVSInstalled, AutoInitialize):
-
-    debug_extra = None
-
-    @classmethod
-    def _initialize(cls) -> None:
-        cls.debug_extra = debug_extra(cls)
-
-    @classmethod
-    def factory(
-        cls, *,
-        msvs_instances,
-    ):
-
-        msvs_instances = sorted(
-            msvs_instances, key=cmp_to_key(MSVSInstance.default_order)
-        )
-
-        msvs_channel_map = {
-            vs_channel_def: {}
-            for vs_channel_def in MSVC.Config.MSVS_CHANNEL_DEFINITION_LIST
-        }
-
-        vs_edition_instances = {}
-        vs_channel_instances = {}
-
-        # channel key: (ANY, None)
-        vs_anychannel_key = _VSKeys.msvs_channel_key(
-            vs_channel_def=MSVC.Config.MSVS_CHANNEL_ANY,
-        )
-
-        for msvs_instance in msvs_instances:
-
-            debug(
-                'msvs instance: id_str=%s, msvc_version=%s, vs_dir=%s, vs_exec=%s',
-                repr(msvs_instance.id_str),
-                repr(msvs_instance.msvc_version),
-                repr(msvs_instance.vs_dir),
-                repr(msvs_instance.vs_executable_norm),
-                extra=cls.debug_extra,
-            )
-
-            # edition key: (product, ANY, None, None)
-            vs_edition_any_key = _VSKeys.msvs_edition_key(
-                vs_product_def=msvs_instance.vs_product_def,
-                vs_channel_def=MSVC.Config.MSVS_CHANNEL_ANY,
-            )
-
-            # edition key: (product, ANY, component, None)
-            vs_edition_anychannel_component_key = _VSKeys.msvs_edition_key(
-                vs_product_def=msvs_instance.vs_product_def,
-                vs_channel_def=MSVC.Config.MSVS_CHANNEL_ANY,
-                vs_componentid_def=msvs_instance.vs_componentid_def,
-            )
-
-            # all editions keys
-            vs_edition_keys = (
-                vs_edition_any_key,
-                msvs_instance.vs_edition_channel_key,
-                vs_edition_anychannel_component_key,
-                msvs_instance.vs_edition_channel_component_key,
-                msvs_instance.vs_edition_channel_component_seqnbr_key,
-            )
-
-            # channel key: (ANY, component)
-            vs_anychannel_component_key = _VSKeys.msvs_channel_key(
-                vs_channel_def=MSVC.Config.MSVS_CHANNEL_ANY,
-                vs_componentid_def=msvs_instance.vs_componentid_def,
-            )
-
-            # all channel keys
-            vs_channel_keys = (
-                vs_anychannel_key,
-                msvs_instance.vs_channel_key,
-                vs_anychannel_component_key,
-                msvs_instance.vs_channel_component_key,
-            )
-
-            for vs_edition_key in vs_edition_keys:
-                vs_edition_instances.setdefault(vs_edition_key, []).append(msvs_instance)
-
-            for vs_channel_key in vs_channel_keys:
-                vs_channel_instances.setdefault(vs_channel_key, []).append(msvs_instance)
-
-            # msvc_version support
-
-            if msvs_instance.msvc_version != msvs_instance.msvc_verstr:
-                versions = [msvs_instance.msvc_verstr, msvs_instance.msvc_version]
-            else:
-                versions = [msvs_instance.msvc_verstr]
-
-            vs_channel_defs = MSVC.Config.MSVS_CHANNEL_MEMBERLISTS[msvs_instance.vs_channel_def]
-
-            for vcver in versions:
-                for vs_channel_def in vs_channel_defs:
-
-                    msvs_version_map = msvs_channel_map[vs_channel_def]
-
-                    msvs_instance_list = msvs_version_map.setdefault(vcver, [])
-                    msvs_instance_list.append(msvs_instance)
-
-        msvs_installed = cls(
-            msvs_instances=msvs_instances,
-            msvs_edition_instances_map=vs_edition_instances,
-            msvs_channel_instances_map=vs_channel_instances,
-            msvs_channel_map=msvs_channel_map,
-        )
-
-        if DEBUG_ENABLED:
-            msvs_installed._debug_dump()
-
-        return msvs_installed
-
-    def _debug_dump(self) -> None:
-
-        for vs_channel_key, msvc_instance_list in self.msvs_channel_instances_map.items():
-            msvc_instances = [msvc_instance.id_str for msvc_instance in msvc_instance_list]
-            debug(
-                'msvs_channel_instances_map[%s]=%s',
-                repr(vs_channel_key.serialize()),
-                repr(msvc_instances),
-                extra=self.debug_extra,
-            )
-
-        for vs_edition_key, msvc_instance_list in self.msvs_edition_instances_map.items():
-            msvc_instances = [msvc_instance.id_str for msvc_instance in msvc_instance_list]
-            debug(
-                'msvs_edition_instances_map[%s]=%s',
-                repr(vs_edition_key.serialize()),
-                repr(msvc_instances),
-                extra=self.debug_extra,
-            )
-
-        for vs_channel_def, msvs_version_map in self.msvs_channel_map.items():
-            for vcver, msvs_instance_list in msvs_version_map.items():
-                msvs_instances = [msvs_instance.id_str for msvs_instance in msvs_instance_list]
-                debug(
-                    'msvs_version_map[%s][%s]=%s',
-                    repr(vs_channel_def.vs_channel_suffix),
-                    repr(vcver),
-                    repr(msvs_instances),
-                    extra=self.debug_extra,
-                )
-
-    def _msvs_instances_internal(
-        self, *,
-        vs_product_def=None,
-        vs_channel_def=None,
-        vs_componentid_def=None,
-        vs_sequence_nbr=None,
-    ):
-
-        if not vs_channel_def:
-            vs_channel_def = _VSChannel.get_default_channel()
-
-        if vs_product_def:
-
-            query_key = _VSKeys.msvs_edition_key(
-                vs_product_def=vs_product_def,
-                vs_channel_def=vs_channel_def,
-                vs_componentid_def=vs_componentid_def,
-                vs_sequence_nbr=vs_sequence_nbr
-            )
-
-            msvs_instances = self.msvs_edition_instances_map.get(query_key, [])
-
-        else:
-
-            query_key = _VSKeys.msvs_channel_key(
-                vs_channel_def=vs_channel_def,
-                vs_componentid_def=vs_componentid_def,
-            )
-
-            msvs_instances = self.msvs_channel_instances_map.get(query_key, [])
-
-        debug(
-            'query_key=%s, n_msvs_instances=%s',
-            repr(query_key.serialize()),
-            repr(len(msvs_instances)),
-            extra=self.debug_extra,
-        )
-
-        return msvs_instances, query_key
-
-_MSVCInstance = namedtuple('_MSVCInstance', [
-    'id_str',
-    'msvs_base',
-    'vc_version_def',
-    'vc_feature_map',
-    'vc_dir',
-    'vc_dir_norm',
-    'is_sdkversion_supported',
-])
-
-class MSVCInstance(_MSVCInstance, AutoInitialize):
-
-    debug_extra = None
-
-    @classmethod
-    def _initialize(cls) -> None:
-        cls.debug_extra = debug_extra(cls)
-
-    @classmethod
-    def factory(
-        cls, *,
-        msvs_base,
-        vc_version_def,
-        vc_feature_map,
-        vc_dir,
-    ):
-
-        id_str = '{}({})'.format(cls.__name__, ', '.join(msvs_base.id_comps))
-
-        if vc_feature_map is None:
-            vc_feature_map = {}
-
-        msvc_instance = cls(
-            id_str=id_str,
-            msvs_base=msvs_base,
-            vc_version_def=vc_version_def,
-            vc_feature_map=vc_feature_map,
-            vc_dir=vc_dir,
-            vc_dir_norm=_Util.normalize_path(vc_dir),
-            is_sdkversion_supported=cls._is_sdkversion_supported(msvs_base),
-        )
-
-        return msvc_instance
-
-    @staticmethod
-    def default_order(a, b):
-        return MSVSBase.default_order(a.msvs_base, b.msvs_base)
-
-    @property
-    def id_comps(self):
-        return self.msvs_base.id_comps
-
-    @property
-    def msvs_instance(self):
-        return self.msvs_base.msvs_instance
-
-    @staticmethod
-    def _is_sdkversion_supported(msvs_base) -> bool:
-
-        vs_componentid_def = msvs_base.vs_component_def.vs_componentid_def
-        vs_product_numeric = msvs_base.vs_product_def.vs_product_numeric
-
-        if vs_product_numeric >= 2017:
-            # VS2017 and later
-            is_supported = True
-        elif vs_product_numeric == 2015:
-            # VS2015:
-            #   False: Express, BuildTools
-            #   True:  Develop, CmdLine
-            if vs_componentid_def == MSVC.Config.MSVS_COMPONENTID_EXPRESS:
-                is_supported = False
-            elif vs_componentid_def == MSVC.Config.MSVS_COMPONENTID_BUILDTOOLS:
-                is_supported = False
-            else:
-                is_supported = True
-        else:
-            # VS2013 and earlier
-            is_supported = False
-
-        return is_supported
-
-    def skip_uwp_target(self, env) -> bool:
-        skip = False
-        if self.vc_feature_map:
-            target_arch = env.get('TARGET_ARCH')
-            uwp_target_is_supported = self.vc_feature_map.get('uwp_target_is_supported', {})
-            is_supported = uwp_target_is_supported.get(target_arch, True)
-            skip = bool(not is_supported)
-        debug(
-            'skip=%s, msvc_version=%s',
-            repr(skip), repr(self.vc_version_def.msvc_version), extra=self.debug_extra
-        )
-        return skip
-
-    def is_uwp_target_supported(self, target_arch=None) -> bool:
-
-        vs_componentid_def = self.msvs_base.vs_component_def.vs_componentid_def
-        vs_product_numeric = self.msvs_base.vs_product_def.vs_product_numeric
-
-        is_target = False
-        if vs_product_numeric >= 2017:
-            # VS2017 and later
-            is_supported = True
-        elif vs_product_numeric == 2015:
-            # VS2015:
-            #   Maybe: Express
-            #   False: BuildTools
-            #   True:  Develop, CmdLine
-            if vs_componentid_def == MSVC.Config.MSVS_COMPONENTID_EXPRESS:
-                uwp_target_is_supported = self.vc_feature_map.get('uwp_target_is_supported', {})
-                is_supported = uwp_target_is_supported.get(target_arch, True)
-                is_target = True
-            elif vs_componentid_def == MSVC.Config.MSVS_COMPONENTID_BUILDTOOLS:
-                is_supported = False
-            else:
-                is_supported = True
-        else:
-            # VS2013 and earlier
-            is_supported = False
-
-        debug(
-            'is_supported=%s, is_target=%s, msvc_version=%s, msvs_component=%s, target_arch=%s',
-            is_supported, is_target,
-            repr(self.vc_version_def.msvc_version),
-            repr(vs_componentid_def.vs_component_id),
-            repr(target_arch),
-            extra=self.debug_extra,
-        )
-
-        return is_supported, is_target
-
-    # convenience properties: reduce member lookup chain for readability
-
-    @property
-    def is_express(self) -> bool:
-        return self.msvs_base.is_express
-
-    @property
-    def is_buildtools(self) -> bool:
-        return self.msvs_base.is_buildtools
-
-    @property
-    def is_vcforpython(self) -> bool:
-        return self.msvs_base.is_vcforpython
-
-    @property
-    def vs_edition_channel_component_seqnbr_key(self):
-        return self.msvs_base.vs_edition_channel_component_seqnbr_key
-
-    @property
-    def vs_edition_channel_component_key(self):
-        return self.msvs_base.vs_edition_channel_component_key
-
-    @property
-    def vs_edition_channel_key(self):
-        return self.msvs_base.vs_edition_channel_key
-
-    @property
-    def vs_channel_component_key(self):
-        return self.msvs_base.vs_channel_component_key
-
-    @property
-    def vs_channel_key(self):
-        return self.msvs_base.vs_channel_key
-
-    @property
-    def msvc_version(self):
-        return self.vc_version_def.msvc_version
-
-    @property
-    def msvc_vernum(self):
-        return self.vc_version_def.msvc_vernum
-
-    @property
-    def msvc_verstr(self):
-        return self.vc_version_def.msvc_verstr
-
-    @property
-    def vs_product_def(self):
-        return self.msvs_base.vs_product_def
-
-    @property
-    def vs_channel_def(self):
-        return self.msvs_base.vs_channel_def
-
-    @property
-    def vs_componentid_def(self):
-        return self.msvs_base.vs_component_def.vs_componentid_def
-
-    @property
-    def vs_product(self):
-        return self.msvs_base.vs_product_def.vs_product
-
-    @property
-    def vs_product_numeric(self):
-        return self.msvs_base.vs_product_def.vs_product_numeric
-
-    @property
-    def vs_channel_id(self):
-        return self.msvs_base.vs_channel_def.vs_channel_id
-
-    @property
-    def vs_component_id(self):
-        return self.msvs_base.vs_component_def.vs_componentid_def.vs_component_id
-
-    @property
-    def vs_sequence_nbr(self):
-        return self.msvs_base.vs_sequence_nbr
-
-_MSVCInstalled = namedtuple('_MSVCInstalled', [
-    'msvc_instances',
-    'msvs_edition_instances_map',
-    'msvs_channel_instances_map',
-    'msvs_channel_map', # TODO(JCB): remove?
-])
-
-class MSVCInstalled(_MSVCInstalled, AutoInitialize):
-
-    debug_extra = None
-
-    @classmethod
-    def _initialize(cls) -> None:
-        cls.debug_extra = debug_extra(cls)
-
-    @classmethod
-    def factory(
-        cls, *,
-        msvc_instances,
-    ):
-
-        msvc_instances = sorted(
-            msvc_instances, key=cmp_to_key(MSVCInstance.default_order)
-        )
-
-        msvs_channel_map = {
-            vs_channel_def: {}
-            for vs_channel_def in MSVC.Config.MSVS_CHANNEL_DEFINITION_LIST
-        }
-
-        vs_edition_instances = {}
-        vs_channel_instances = {}
-
-        # channel key: (ANY, None)
-        vs_anychannel_key = _VSKeys.msvs_channel_key(
-            vs_channel_def=MSVC.Config.MSVS_CHANNEL_ANY,
-        )
-
-        for msvc_instance in msvc_instances:
-
-            debug(
-                'msvc_instance: id_str=%s, msvc_version=%s, vc_dir=%s',
-                repr(msvc_instance.id_str),
-                repr(msvc_instance.msvc_version),
-                repr(msvc_instance.vc_dir),
-                extra=cls.debug_extra,
-            )
-
-            # edition key: (product, ANY, None, None)
-            vs_edition_any_key = _VSKeys.msvs_edition_key(
-                vs_product_def=msvc_instance.vs_product_def,
-                vs_channel_def=MSVC.Config.MSVS_CHANNEL_ANY,
-            )
-
-            # edition key: (product, ANY, component, None)
-            vs_edition_anychannel_component_key = _VSKeys.msvs_edition_key(
-                vs_product_def=msvc_instance.vs_product_def,
-                vs_channel_def=MSVC.Config.MSVS_CHANNEL_ANY,
-                vs_componentid_def=msvc_instance.vs_componentid_def,
-            )
-
-            # all editions keys
-            vs_edition_keys = (
-                vs_edition_any_key,
-                msvc_instance.vs_edition_channel_key,
-                vs_edition_anychannel_component_key,
-                msvc_instance.vs_edition_channel_component_key,
-                msvc_instance.vs_edition_channel_component_seqnbr_key,
-            )
-
-            # channel key: (ANY, component)
-            vs_anychannel_component_key = _VSKeys.msvs_channel_key(
-                vs_channel_def=MSVC.Config.MSVS_CHANNEL_ANY,
-                vs_componentid_def=msvc_instance.vs_componentid_def,
-            )
-
-            # all channel keys
-            vs_channel_keys = (
-                vs_anychannel_key,
-                msvc_instance.vs_channel_key,
-                vs_anychannel_component_key,
-                msvc_instance.vs_channel_component_key,
-            )
-
-            for vs_edition_key in vs_edition_keys:
-                vs_edition_instances.setdefault(vs_edition_key, []).append(msvc_instance)
-
-            for vs_channel_key in vs_channel_keys:
-                vs_channel_instances.setdefault(vs_channel_key, []).append(msvc_instance)
-
-            # msvc_version support
-
-            if msvc_instance.msvc_version != msvc_instance.msvc_verstr:
-                versions = [msvc_instance.msvc_verstr, msvc_instance.msvc_version]
-            else:
-                versions = [msvc_instance.msvc_verstr]
-
-            vs_channel_defs = MSVC.Config.MSVS_CHANNEL_MEMBERLISTS[msvc_instance.vs_channel_def]
-
-            for vcver in versions:
-
-                for vs_channel_def in vs_channel_defs:
-
-                    msvc_version_map = msvs_channel_map[vs_channel_def]
-
-                    msvc_instance_list = msvc_version_map.setdefault(vcver, [])
-                    msvc_instance_list.append(msvc_instance)
-
-        msvc_installed = cls(
-            msvc_instances=msvc_instances,
-            msvs_edition_instances_map=vs_edition_instances,
-            msvs_channel_instances_map=vs_channel_instances,
-            msvs_channel_map=msvs_channel_map,
-        )
-
-        if DEBUG_ENABLED:
-            msvc_installed._debug_dump()
-
-        return msvc_installed
-
-    def _debug_dump(self) -> None:
-
-        for vs_channel_key, msvc_instance_list in self.msvs_channel_instances_map.items():
-            msvc_instances = [msvc_instance.id_str for msvc_instance in msvc_instance_list]
-            debug(
-                'msvs_channel_instances_map[%s]=%s',
-                repr(vs_channel_key.serialize()),
-                repr(msvc_instances),
-                extra=self.debug_extra,
-            )
-
-        for vs_edition_key, msvc_instance_list in self.msvs_edition_instances_map.items():
-            msvc_instances = [msvc_instance.id_str for msvc_instance in msvc_instance_list]
-            debug(
-                'msvs_edition_instances_map[%s]=%s',
-                repr(vs_edition_key.serialize()),
-                repr(msvc_instances),
-                extra=self.debug_extra,
-            )
-
-        for vs_channel_def, msvc_version_map in self.msvs_channel_map.items():
-            for vcver, msvc_instance_list in msvc_version_map.items():
-                msvc_instances = [msvc_instance.id_str for msvc_instance in msvc_instance_list]
-                debug(
-                    'msvc_version_map[%s][%s]=%s',
-                    repr(vs_channel_def.vs_channel_suffix),
-                    repr(vcver),
-                    repr(msvc_instances),
-                    extra=self.debug_extra,
-                )
-
-    def _msvc_instances_internal(
-        self, *,
-        vs_product_def=None,
-        vs_channel_def=None,
-        vs_componentid_def=None,
-        vs_sequence_nbr=None,
-    ):
-
-        if not vs_channel_def:
-            vs_channel_def = _VSChannel.get_default_channel()
-
-        if vs_product_def:
-
-            query_key = _VSKeys.msvs_edition_key(
-                vs_product_def=vs_product_def,
-                vs_channel_def=vs_channel_def,
-                vs_componentid_def=vs_componentid_def,
-                vs_sequence_nbr=vs_sequence_nbr
-            )
-
-            msvc_instances = self.msvs_edition_instances_map.get(query_key, [])
-
-        else:
-
-            query_key = _VSKeys.msvs_channel_key(
-                vs_channel_def=vs_channel_def,
-                vs_componentid_def=vs_componentid_def,
-            )
-
-            msvc_instances = self.msvs_channel_instances_map.get(query_key, [])
-
-        debug(
-            'query_key=%s, n_msvc_instances=%s',
-            repr(query_key.serialize()),
-            repr(len(msvc_instances)),
-            extra=self.debug_extra,
-        )
-
-        return msvc_instances, query_key
-
-_MSVSManager = namedtuple('_MSVSManager', [
-    'msvc_installed',
-    'msvs_installed',
-])
-
-class MSVSManager(_MSVSManager, AutoInitialize):
-
-    debug_extra = None
-
-    @classmethod
-    def _initialize(cls) -> None:
-        cls.debug_extra = debug_extra(cls)
-
-    @classmethod
-    def factory(cls, msvc_installed, msvs_installed):
-
-        msvs_manager = cls(
-            msvc_installed=msvc_installed,
-            msvs_installed=msvs_installed,
-        )
-
-        debug(
-            'n_msvc_instances=%d, n_msvs_instances=%d',
-            len(msvc_installed.msvc_instances),
-            len(msvs_installed.msvs_instances),
-            extra=cls.debug_extra,
-        )
-
-        return msvs_manager
-
-    def query_msvs_instances(
-        self, *,
-        msvc_version,
-    ):
-
-        # TODO(JCB): temporary placeholder
-
-        prefix, suffix = MSVC.Util.get_msvc_version_prefix_suffix(msvc_version)
-
-        vs_product_def = MSVC.Config.MSVC_VERSION_INTERNAL[prefix]
-
-        vs_channel_def = None
-
-        if suffix == 'Exp':
-            vs_componentid_def = MSVC.Config.MSVS_COMPONENTID_EXPRESS
-        else:
-            vs_componentid_def = None
-
-        msvs_instances, query_key = self.msvs_installed._msvs_instances_internal(
-            vs_product_def=vs_product_def,
-            vs_channel_def=vs_channel_def,
-            vs_componentid_def=vs_componentid_def,
-        )
-
-        return msvs_instances, query_key
-
-    def query_msvc_instances(
-        self, *,
-        msvc_version,
-    ):
-
-        # TODO(JCB): temporary placeholder
-
-        prefix, suffix = MSVC.Util.get_msvc_version_prefix_suffix(msvc_version)
-
-        vs_product_def = MSVC.Config.MSVC_VERSION_INTERNAL[prefix]
-
-        vs_channel_def = None
-
-        if suffix == 'Exp':
-            vs_componentid_def = MSVC.Config.MSVS_COMPONENTID_EXPRESS
-        else:
-            vs_componentid_def = None
-
-        msvc_instances, query_key = self.msvc_installed._msvc_instances_internal(
-            vs_product_def=vs_product_def,
-            vs_channel_def=vs_channel_def,
-            vs_componentid_def=vs_componentid_def,
-        )
-
-        return msvc_instances, query_key
-
-class _VSDetectCommon(AutoInitialize):
-
-    debug_extra = None
-
-    @classmethod
-    def _initialize(cls) -> None:
-        cls.debug_extra = debug_extra(cls)
-
-    VSDetectedBinaries = namedtuple("VSDetectedBinaries", [
-        'bitfields',  # detect values
-        'have_dev',  # develop ide binary
-        'have_exp',  # express ide binary
-        'have_exp_win',  # express windows ide binary
-        'have_exp_web',  # express web ide binary
-        'vs_exec',  # vs binary
-    ])
-
-    @classmethod
-    def msvs_detect(cls, vs_dir, vs_cfg):
-
-        vs_exec = None
-        vs_script = None
-
-        vs_binary_cfg = vs_cfg.vs_binary_cfg
-        vs_batch_cfg = vs_cfg.vs_batch_cfg
-
-        ide_path = os.path.join(vs_dir, vs_binary_cfg.pathcomp)
-
-        bitfields = 0b_0000
-        for ide_program in vs_binary_cfg.programs:
-            prog = os.path.join(ide_path, ide_program.program)
-            if not os.path.exists(prog):
-                continue
-            bitfields |= ide_program.bitfield
-            if vs_exec:
-                continue
-            if not ide_program.bitfield & _VSConfig.EXECUTABLE_MASK:
-                continue
-            vs_exec = prog
-
-        have_dev = bool(bitfields & _VSConfig.BITFIELD_DEVELOP)
-        have_exp = bool(bitfields & _VSConfig.BITFIELD_EXPRESS)
-        have_exp_win = bool(bitfields & _VSConfig.BITFIELD_EXPRESS_WIN)
-        have_exp_web = bool(bitfields & _VSConfig.BITFIELD_EXPRESS_WEB)
-
-        binaries_t = cls.VSDetectedBinaries(
-            bitfields=bitfields,
-            have_dev=have_dev,
-            have_exp=have_exp,
-            have_exp_win=have_exp_win,
-            have_exp_web=have_exp_web,
-            vs_exec=vs_exec,
-        )
-
-        script_file = os.path.join(vs_dir, vs_batch_cfg.pathcomp, vs_batch_cfg.script)
-        if os.path.exists(script_file):
-            vs_script = script_file
-
-        debug(
-            'vs_dir=%s, dev=%s, exp=%s, exp_win=%s, exp_web=%s, vs_exec=%s, vs_script=%s',
-            repr(vs_dir),
-            binaries_t.have_dev, binaries_t.have_exp,
-            binaries_t.have_exp_win, binaries_t.have_exp_web,
-            repr(binaries_t.vs_exec),
-            repr(vs_script),
-            extra=cls.debug_extra,
-        )
-
-        return binaries_t, vs_script
-
-class _VSDetectVSWhere(AutoInitialize):
-
-    DETECT_CONFIG = _VSConfig.DETECT_VSWHERE
-
-    # initialization
-
-    debug_extra = None
-    reset_funcs = []
-
-    @classmethod
-    def _initialize(cls) -> None:
-        cls.debug_extra = debug_extra(cls)
-        cls.reset()
-
-    @classmethod
-    def reset(cls) -> None:
-
-        cls.vswhere_executable_seen = set()
-        cls.vswhere_executables = []
-
-        cls.vs_dir_seen = set()
-
-        cls.msvs_sequence_nbr = {}
-
-        cls.msvs_instances = []
-        cls.msvc_instances = []
-
-    @classmethod
-    def register_reset_func(cls, func) -> None:
-        if func:
-            cls.reset_funcs.append(func)
-
-    @classmethod
-    def call_reset_funcs(cls) -> None:
-        for func in cls.reset_funcs:
-            func()
-
-    @classmethod
-    def _filter_vswhere_paths(cls, vswhere_env):
-        vswhere_executables = _VSWhere.find_executables(vswhere_env)
-        vswhere_paths = [
-            vswhere_exec.norm
-            for vswhere_exec in vswhere_executables
-            if vswhere_exec.norm not in cls.vswhere_executable_seen
-        ]
-        debug('vswhere_paths=%s', vswhere_paths, extra=cls.debug_extra)
-        return vswhere_paths
-
-    @classmethod
-    def _vswhere_query_json_output(cls, vswhere_exe, vswhere_args):
-
-        vswhere_json = None
-
-        once = True
-        while once:
-            once = False
-            # using break for single exit (unless exception)
-
-            vswhere_cmd = [vswhere_exe] + vswhere_args + ['-format', 'json', '-utf8']
-            debug("running: %s", vswhere_cmd, extra=cls.debug_extra)
-
-            try:
-                cp = subprocess.run(
-                    vswhere_cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, check=False
-                )
-            except OSError as e:
-                errmsg = str(e)
-                debug("%s: %s", type(e).__name__, errmsg, extra=cls.debug_extra)
-                break
-            except Exception as e:
-                errmsg = str(e)
-                debug("%s: %s", type(e).__name__, errmsg, extra=cls.debug_extra)
-                raise
-
-            if not cp.stdout:
-                debug("no vswhere information returned", extra=cls.debug_extra)
-                break
-
-            vswhere_output = cp.stdout.decode('utf8', errors='replace')
-            if not vswhere_output:
-                debug("no vswhere information output", extra=cls.debug_extra)
-                break
-
-            try:
-                vswhere_output_json = json.loads(vswhere_output)
-            except json.decoder.JSONDecodeError:
-                debug("json decode exception loading vswhere output", extra=cls.debug_extra)
-                break
-
-            vswhere_json = vswhere_output_json
-            break
-
-        debug(
-            'vswhere_json=%s, vswhere_exe=%s',
-            bool(vswhere_json), repr(vswhere_exe), extra=cls.debug_extra
-        )
-
-        return vswhere_json
-
-    @classmethod
-    def _msvc_resolve(cls, vc_dir, vs_product_def):
-
-        detect_cfg = cls.DETECT_CONFIG[vs_product_def.vs_product]
-
-        vs_cfg = detect_cfg.vs_cfg
-        vs_dir = os.path.normpath(os.path.join(vc_dir, vs_cfg.root))
-
-        binaries_t, vs_script = _VSDetectCommon.msvs_detect(vs_dir, vs_cfg)
-
-        return binaries_t.vs_exec, vs_script
-
-    @classmethod
-    def _msvc_instance_toolsets(cls, msvc_instance):
-        # TODO(JCB): load toolsets/tools
-        rval = _msvc_instance_check_files_exist(msvc_instance)
-        return rval
-
-    @classmethod
-    def detect(cls, vswhere_env):
-
-        num_instances = len(cls.msvc_instances)
-        num_new_instances = 0
-
-        vswhere_paths = cls._filter_vswhere_paths(vswhere_env)
-        if vswhere_paths:
-
-            num_beg_instances = num_instances
-
-            for vswhere_exe in vswhere_paths:
-
-                if vswhere_exe in cls.vswhere_executable_seen:
-                    continue
-
-                cls.vswhere_executable_seen.add(vswhere_exe)
-                cls.vswhere_executables.append(vswhere_exe)
-
-                debug('vswhere_exe=%s', repr(vswhere_exe), extra=cls.debug_extra)
-
-                vswhere_json = cls._vswhere_query_json_output(
-                    vswhere_exe,
-                    ['-all', '-products', '*', '-prerelease']
-                )
-
-                if not vswhere_json:
-                    continue
-
-                for instance in vswhere_json:
-
-                    #print(json.dumps(instance, indent=4, sort_keys=True))
-
-                    vs_dir = instance.get('installationPath')
-                    if not vs_dir or not _Util.path_exists(vs_dir):
-                        continue
-
-                    vs_norm = _Util.normalize_path(vs_dir)
-                    if vs_norm in cls.vs_dir_seen:
-                        continue
-
-                    vs_version = instance.get('installationVersion')
-                    if not vs_version:
-                        continue
-
-                    product_id = instance.get('productId')
-                    if not product_id:
-                        continue
-
-                    # consider msvs root evaluated at this point
-                    cls.vs_dir_seen.add(vs_norm)
-
-                    vc_dir = os.path.join(vs_dir, 'VC')
-                    if not _Util.path_exists(vc_dir):
-                        continue
-
-                    vs_major = vs_version.split('.')[0]
-                    if vs_major not in MSVC.Config.MSVS_VERSION_MAJOR_MAP:
-                        debug('ignore vs_major: %s', vs_major, extra=cls.debug_extra)
-                        continue
-
-                    vs_product_def = MSVC.Config.MSVS_VERSION_MAJOR_MAP[vs_major]
-
-                    component_id = product_id.split('.')[-1]
-                    vs_component_def = MSVC.Config.VSWHERE_COMPONENT_INTERNAL.get(component_id)
-                    if not vs_component_def:
-                        debug('ignore component_id: %s', component_id, extra=cls.debug_extra)
-                        continue
-
-                    is_prerelease = True if instance.get('isPrerelease', False) else False
-                    if is_prerelease:
-                        vs_channel_def = MSVC.Config.MSVS_CHANNEL_PREVIEW
-                    else:
-                        vs_channel_def = MSVC.Config.MSVS_CHANNEL_RELEASE
-
-                    vs_exec, vs_script = cls._msvc_resolve(vc_dir, vs_product_def)
-
-                    vs_sequence_key = (vs_product_def, vs_channel_def, vs_component_def)
-
-                    cls.msvs_sequence_nbr.setdefault(vs_sequence_key, 0)
-                    cls.msvs_sequence_nbr[vs_sequence_key] += 1
-
-                    vs_sequence_nbr = cls.msvs_sequence_nbr[vs_sequence_key]
-
-                    msvs_base = MSVSBase.factory(
-                        vs_product_def=vs_product_def,
-                        vs_channel_def=vs_channel_def,
-                        vs_component_def=vs_component_def,
-                        vs_sequence_nbr=vs_sequence_nbr,
-                        vs_dir=vs_dir,
-                        vs_version=vs_version,
-                    )
-
-                    vc_version = vs_product_def.vc_buildtools_def.vc_version
-
-                    if msvs_base.is_express:
-                        vc_version += msvs_base.vs_component_suffix
-
-                    vc_version_def = MSVC.Util.msvc_version_components(vc_version)
-
-                    msvc_instance = MSVCInstance.factory(
-                        msvs_base=msvs_base,
-                        vc_version_def=vc_version_def,
-                        vc_feature_map=None,
-                        vc_dir=vc_dir,
-                    )
-
-                    if not cls._msvc_instance_toolsets(msvc_instance):
-                        # no compilers detected don't register objects
-                        continue
-
-                    if vs_exec:
-
-                        # TODO(JCB): register iff compilers and executable?
-                        msvs_instance = MSVSInstance.factory(
-                            msvs_base=msvs_base,
-                            vs_executable=vs_exec,
-                            vs_script=vs_script,
-                            vc_version_def=vc_version_def,
-                        )
-
-                        msvs_base.register_msvs_instance(msvs_instance)
-                        cls.msvs_instances.append(msvs_instance)
-
-                        debug(
-                            'msvs_instance=%s',
-                            repr(msvs_instance.id_str),
-                            extra=cls.debug_extra,
-                        )
-
-                    msvs_base.register_msvc_instance(msvc_instance)
-                    cls.msvc_instances.append(msvc_instance)
-
-                    debug(
-                        'msvc_instance=%s',
-                        repr(msvc_instance.id_str),
-                        extra=cls.debug_extra,
-                    )
-
-            num_instances = len(cls.msvc_instances)
-            num_new_instances = num_instances - num_beg_instances
-
-        if num_new_instances > 0:
-            cls.call_reset_funcs()
-            debug(
-                'num_new_instances=%s, num_instances=%s',
-                num_new_instances, num_instances, extra=cls.debug_extra
-            )
-
-        return num_new_instances
-
-class _VSDetectRegistry(AutoInitialize):
-
-    DETECT_CONFIG = _VSConfig.DETECT_REGISTRY
-
-    # initialization
-
-    debug_extra = None
-
-    @classmethod
-    def _initialize(cls) -> None:
-        cls.debug_extra = debug_extra(cls)
-        cls.reset()
-
-    @classmethod
-    def reset(cls) -> None:
-
-        cls.registry_once = False
-
-        cls.vc_dir_seen = set()
-
-        cls.msvs_sequence_nbr = {}
-
-        cls.msvs_instances = []
-        cls.msvc_instances = []
-
-    # VS2015 buildtools batch file call detection
-    #    vs2015 buildtools do not support sdk_version or UWP arguments
-
-    _VS2015BT_PATH = r'..\Microsoft Visual C++ Build Tools\vcbuildtools.bat'
-
-    _VS2015BT_REGEX_STR = ''.join([
-        r'^\s*if\s+exist\s+',
-        re.escape(fr'"%~dp0..\{_VS2015BT_PATH}"'),
-        r'\s+goto\s+setup_buildsku\s*$',
-    ])
-
-    _VS2015BT_VCVARS_BUILDTOOLS = re.compile(_VS2015BT_REGEX_STR, re.IGNORECASE)
-    _VS2015BT_VCVARS_STOP = re.compile(r'^\s*[:]Setup_VS\s*$', re.IGNORECASE)
-
-    @classmethod
-    def _vs_buildtools_2015_vcvars(cls, vcvars_file) -> bool:
-        have_buildtools_vcvars = False
-        with open(vcvars_file) as fh:
-            for line in fh:
-                if cls._VS2015BT_VCVARS_BUILDTOOLS.match(line):
-                    have_buildtools_vcvars = True
-                    break
-                if cls._VS2015BT_VCVARS_STOP.match(line):
-                    break
-        return have_buildtools_vcvars
-
-    @classmethod
-    def _vs_buildtools_2015(cls, vs_dir, vc_dir) -> bool:
-
-        is_buildtools = False
-
-        do_once = True
-        while do_once:
-            do_once = False
-
-            buildtools_file = os.path.join(vs_dir, cls._VS2015BT_PATH)
-            have_buildtools = os.path.exists(buildtools_file)
-            debug('have_buildtools=%s', have_buildtools, extra=cls.debug_extra)
-            if not have_buildtools:
-                break
-
-            vcvars_file = os.path.join(vc_dir, 'vcvarsall.bat')
-            have_vcvars = os.path.exists(vcvars_file)
-            debug('have_vcvars=%s', have_vcvars, extra=cls.debug_extra)
-            if not have_vcvars:
-                break
-
-            have_buildtools_vcvars = cls._vs_buildtools_2015_vcvars(vcvars_file)
-            debug('have_buildtools_vcvars=%s', have_buildtools_vcvars, extra=cls.debug_extra)
-            if not have_buildtools_vcvars:
-                break
-
-            is_buildtools = True
-
-        debug('is_buildtools=%s', is_buildtools, extra=cls.debug_extra)
-        return is_buildtools
-
-    _VS2015EXP_VCVARS_LIBPATH = re.compile(
-        ''.join([
-            r'^\s*\@if\s+exist\s+\"\%VCINSTALLDIR\%LIB\\store\\(amd64|arm)"\s+',
-            r'set (LIB|LIBPATH)=\%VCINSTALLDIR\%LIB\\store\\(amd64|arm);.*\%(LIB|LIBPATH)\%\s*$'
-        ]),
-        re.IGNORECASE
-    )
-
-    _VS2015EXP_VCVARS_STOP = re.compile(r'^\s*[:]GetVSCommonToolsDir\s*$', re.IGNORECASE)
-
-    @classmethod
-    def _vs_express_2015_vcvars(cls, vcvars_file) -> bool:
-        n_libpath = 0
-        with open(vcvars_file) as fh:
-            for line in fh:
-                if cls._VS2015EXP_VCVARS_LIBPATH.match(line):
-                    n_libpath += 1
-                elif cls._VS2015EXP_VCVARS_STOP.match(line):
-                    break
-        have_uwp_fix = bool(n_libpath >= 2)
-        return have_uwp_fix
-
-    @classmethod
-    def _vs_express_2015(cls, vc_dir):
-
-        have_uwp_amd64 = False
-        have_uwp_arm = False
-
-        vcvars_file = os.path.join(vc_dir, r'vcvarsall.bat')
-        if os.path.exists(vcvars_file):
-
-            vcvars_file = os.path.join(vc_dir, r'bin\x86_amd64\vcvarsx86_amd64.bat')
-            if os.path.exists(vcvars_file):
-                have_uwp_fix = cls._vs_express_2015_vcvars(vcvars_file)
-                if have_uwp_fix:
-                    have_uwp_amd64 = True
-
-            vcvars_file = os.path.join(vc_dir, r'bin\x86_arm\vcvarsx86_arm.bat')
-            if os.path.exists(vcvars_file):
-                have_uwp_fix = cls._vs_express_2015_vcvars(vcvars_file)
-                if have_uwp_fix:
-                    have_uwp_arm = True
-
-        debug('have_uwp_amd64=%s, have_uwp_arm=%s', have_uwp_amd64, have_uwp_arm, extra=cls.debug_extra)
-        return have_uwp_amd64, have_uwp_arm
-
-    # winsdk installed 2010 [7.1], 2008 [7.0, 6.1] folders
-
-    _REGISTRY_WINSDK_VERSIONS = {'10.0', '9.0'}
-
-    @classmethod
-    def _msvc_dir_is_winsdk_only(cls, vc_dir, msvc_version) -> bool:
-
-        # detect winsdk-only installations
-        #
-        #     registry keys:
-        #         [prefix]\VisualStudio\SxS\VS7\10.0  <undefined>
-        #         [prefix]\VisualStudio\SxS\VC7\10.0  product directory
-        #         [prefix]\VisualStudio\SxS\VS7\9.0   <undefined>
-        #         [prefix]\VisualStudio\SxS\VC7\9.0   product directory
-        #
-        #     winsdk notes:
-        #         - winsdk installs do not define the common tools env var
-        #         - the product dir is detected but the vcvars batch files will fail
-        #         - regular installations populate the VS7 registry keys
-
-        if msvc_version not in cls._REGISTRY_WINSDK_VERSIONS:
-
-            is_sdk = False
-
-            debug('is_sdk=%s, msvc_version=%s', is_sdk, repr(msvc_version), extra=cls.debug_extra)
-
-        else:
-
-            vc_suffix = MSVC.Registry.vstudio_sxs_vc7(msvc_version)
-            vc_qresults = [record[0] for record in MSVC.Registry.microsoft_query_paths(vc_suffix)]
-            vc_regdir = vc_qresults[0] if vc_qresults else None
-
-            if vc_regdir != vc_dir:
-                # registry vc path is not the current vc dir
-
-                is_sdk = False
-
-                debug(
-                    'is_sdk=%s, msvc_version=%s, vc_dir=%s, vc_regdir=%s',
-                    is_sdk, repr(msvc_version), repr(vc_dir), repr(vc_regdir), extra=cls.debug_extra
-                )
-
-            else:
-                # registry vc dir is the current vc root
-
-                vs_suffix = MSVC.Registry.vstudio_sxs_vs7(msvc_version)
-                vs_qresults = [record[0] for record in MSVC.Registry.microsoft_query_paths(vs_suffix)]
-                vs_dir = vs_qresults[0] if vs_qresults else None
-
-                is_sdk = bool(not vs_dir and vc_dir)
-
-                debug(
-                    'is_sdk=%s, msvc_version=%s, vc_dir=%s, vs_dir=%s',
-                    is_sdk, repr(msvc_version), repr(vc_dir), repr(vs_dir), extra=cls.debug_extra
-                )
-
-        return is_sdk
-
-    @classmethod
-    def _msvc_resolve(cls, vc_dir, vs_product_def, is_vcforpython, vs_version):
-
-        detect_cfg = cls.DETECT_CONFIG[vs_product_def.vs_product]
-
-        vc_feature_map = {}
-
-        vs_cfg = detect_cfg.vs_cfg
-        vs_dir = os.path.normpath(os.path.join(vc_dir, vs_cfg.root))
-
-        binaries_t, vs_script = _VSDetectCommon.msvs_detect(vs_dir, vs_cfg)
-
-        vs_product_numeric = vs_product_def.vs_product_numeric
-        vc_version = vs_product_def.vc_buildtools_def.vc_version
-
-        if binaries_t.have_dev:
-            vs_component_def = MSVC.Config.REGISTRY_COMPONENT_DEVELOP
-        elif binaries_t.have_exp:
-            vs_component_def = MSVC.Config.REGISTRY_COMPONENT_EXPRESS
-        elif vs_product_numeric == 2008 and is_vcforpython:
-            vs_component_def = MSVC.Config.REGISTRY_COMPONENT_PYTHON
-        elif binaries_t.have_exp_win:
-            vs_component_def = None
-        elif binaries_t.have_exp_web:
-            vs_component_def = None
-        elif cls._msvc_dir_is_winsdk_only(vc_dir, vc_version):
-            vs_component_def = None
-        else:
-            vs_component_def = MSVC.Config.REGISTRY_COMPONENT_CMDLINE
-
-        if vs_component_def and vs_product_numeric == 2015:
-
-            # VS2015:
-            #     remap DEVELOP => ENTERPRISE, PROFESSIONAL, COMMUNITY
-            #     remap CMDLINE => BUILDTOOLS [conditional]
-            #     process EXPRESS
-
-            if vs_component_def == MSVC.Config.REGISTRY_COMPONENT_DEVELOP:
-
-                for reg_component, reg_component_def in [
-                    ('community', MSVC.Config.REGISTRY_COMPONENT_COMMUNITY),
-                    ('professional', MSVC.Config.REGISTRY_COMPONENT_PROFESSIONAL),
-                    ('enterprise', MSVC.Config.REGISTRY_COMPONENT_ENTERPRISE),
-                ]:
-                    suffix = MSVC.Registry.devdiv_vs_servicing_component(vs_version, reg_component)
-                    qresults = MSVC.Registry.microsoft_query_keys(suffix, usrval=reg_component_def)
-                    if not qresults:
-                        continue
-                    vs_component_def = qresults[0][-1]
-                    break
-
-            elif vs_component_def == MSVC.Config.REGISTRY_COMPONENT_CMDLINE:
-
-                if cls._vs_buildtools_2015(vs_dir, vc_dir):
-                    vs_component_def = MSVC.Config.REGISTRY_COMPONENT_BUILDTOOLS
-
-            elif vs_component_def == MSVC.Config.REGISTRY_COMPONENT_EXPRESS:
-
-                have_uwp_amd64, have_uwp_arm = cls._vs_express_2015(vc_dir)
-
-                uwp_target_is_supported = {
-                    'x86': True,
-                    'amd64': have_uwp_amd64,
-                    'arm': have_uwp_arm,
-                }
-
-                vc_feature_map['uwp_target_is_supported'] = uwp_target_is_supported
-
-        debug(
-            'vs_product=%s, vs_component=%s, vc_dir=%s, vc_feature_map=%s',
-            repr(vs_product_numeric),
-            repr(vs_component_def.vs_componentid_def.vs_component_id) if vs_component_def else None,
-            repr(vc_dir),
-            repr(vc_feature_map),
-            extra=cls.debug_extra
-        )
-
-        return vs_component_def, vs_dir, binaries_t.vs_exec, vs_script, vc_feature_map
-
-    @classmethod
-    def _msvc_instance_toolsets(cls, msvc_instance):
-        # TODO(JCB): load toolsets/tools
-        rval = _msvc_instance_check_files_exist(msvc_instance)
-        return rval
-
-    @classmethod
-    def detect(cls):
-
-        num_instances = len(cls.msvc_instances)
-        num_new_instances = 0
-
-        if not cls.registry_once:
-            cls.registry_once = True
-
-            num_beg_instances = num_instances
-
-            is_win64 = common.is_win64()
-
-            for vs_product, config in cls.DETECT_CONFIG.items():
-
-                key_prefix = 'Software\\'
-                for regkey in config.vc_cfg.regkeys:
-
-                    if not regkey.hkroot or not regkey.key:
-                        continue
-
-                    if is_win64:
-                        mskeys = [key_prefix + 'Wow6432Node\\' + regkey.key, key_prefix + regkey.key]
-                    else:
-                        mskeys = [key_prefix + regkey.key]
-
-                    vc_dir = None
-                    for mskey in mskeys:
-                        debug('trying VC registry key %s', repr(mskey), extra=cls.debug_extra)
-                        try:
-                            vc_dir = common.read_reg(mskey, regkey.hkroot)
-                        except OSError:
-                            continue
-                        if vc_dir:
-                            break
-
-                    if not vc_dir:
-                        debug('no VC registry key %s', repr(regkey.key), extra=cls.debug_extra)
-                        continue
-
-                    if vc_dir in cls.vc_dir_seen:
-                        continue
-                    cls.vc_dir_seen.add(vc_dir)
-
-                    if not os.path.exists(vc_dir):
-                        debug(
-                            'reg says vc dir is %s, but it does not exist. (ignoring)',
-                            repr(vc_dir), extra=cls.debug_extra
-                        )
-                        continue
-
-                    if regkey.is_vsroot:
-
-                        vc_dir = os.path.join(vc_dir, 'VC')
-                        debug('convert vs dir to vc dir: %s', repr(vc_dir), extra=cls.debug_extra)
-
-                        if vc_dir in cls.vc_dir_seen:
-                            continue
-                        cls.vc_dir_seen.add(vc_dir)
-
-                        if not os.path.exists(vc_dir):
-                            debug('vc dir does not exist. (ignoring)', repr(vc_dir), extra=cls.debug_extra)
-                            continue
-
-                    vc_norm = _Util.normalize_path(vc_dir)
-                    if vc_norm in cls.vc_dir_seen:
-                        continue
-                    cls.vc_dir_seen.add(vc_norm)
-
-                    vs_product_def = MSVC.Config.MSVS_VERSION_INTERNAL[vs_product]
-
-                    vs_component_def, vs_dir, vs_exec, vs_script, vc_feature_map = cls._msvc_resolve(
-                        vc_norm,
-                        vs_product_def,
-                        regkey.is_vcforpython,
-                        vs_product_def.vs_version
-                    )
-
-                    if not vs_component_def:
-                        continue
-
-                    vs_channel_def = MSVC.Config.MSVS_CHANNEL_RELEASE
-
-                    vs_sequence_key = (vs_product_def, vs_channel_def, vs_component_def)
-
-                    cls.msvs_sequence_nbr.setdefault(vs_sequence_key, 0)
-                    cls.msvs_sequence_nbr[vs_sequence_key] += 1
-
-                    vs_sequence_nbr = cls.msvs_sequence_nbr[vs_sequence_key]
-
-                    msvs_base = MSVSBase.factory(
-                        vs_product_def=vs_product_def,
-                        vs_channel_def=vs_channel_def,
-                        vs_component_def=vs_component_def,
-                        vs_sequence_nbr=vs_sequence_nbr,
-                        vs_dir=vs_dir,
-                        vs_version=vs_product_def.vs_version,
-                    )
-
-                    vc_version = vs_product_def.vc_buildtools_def.vc_version
-
-                    if msvs_base.is_express:
-                        vc_version += msvs_base.vs_component_suffix
-
-                    vc_version_def = MSVC.Util.msvc_version_components(vc_version)
-
-                    msvc_instance = MSVCInstance.factory(
-                        msvs_base=msvs_base,
-                        vc_version_def=vc_version_def,
-                        vc_feature_map=vc_feature_map,
-                        vc_dir=vc_dir,
-                    )
-
-                    if not cls._msvc_instance_toolsets(msvc_instance):
-                        # no compilers detected don't register objects
-                        continue
-
-                    if vs_exec:
-
-                        # TODO(JCB): register iff compilers and executable?
-                        msvs_instance = MSVSInstance.factory(
-                            msvs_base=msvs_base,
-                            vs_executable=vs_exec,
-                            vs_script=vs_script,
-                            vc_version_def=vc_version_def,
-                        )
-
-                        msvs_base.register_msvs_instance(msvs_instance)
-                        cls.msvs_instances.append(msvs_instance)
-
-                        debug(
-                            'msvs_instance=%s',
-                            repr(msvs_instance.id_str),
-                            extra=cls.debug_extra,
-                        )
-
-                    msvs_base.register_msvc_instance(msvc_instance)
-                    cls.msvc_instances.append(msvc_instance)
-
-                    debug(
-                        'msvc_instance=%s',
-                        repr(msvc_instance.id_str),
-                        extra=cls.debug_extra,
-                    )
-
-            num_instances = len(cls.msvc_instances)
-            num_new_instances = num_instances - num_beg_instances
-
-        return num_new_instances
-
-class _VSDetect(AutoInitialize):
-
-    debug_extra = None
-
-    @classmethod
-    def _initialize(cls) -> None:
-        cls.debug_extra = debug_extra(cls)
-        cls._reset()
-
-    @classmethod
-    def _reset(cls) -> None:
-        # volatile: reconstructed when new instances detected
-        cls.msvs_manager = None
-
-    @classmethod
-    def reset(cls) -> None:
-        cls._reset()
-        _VSDetectVSWhere.reset()
-        _VSDetectRegistry.reset()
-
-    @classmethod
-    def register_reset_func(cls, func) -> None:
-        _VSDetectVSWhere.register_reset_func(func)
-
-    @classmethod
-    def detect(cls, vswhere_env):
-
-        num_new_instances = 0
-        num_new_instances += _VSDetectVSWhere.detect(vswhere_env)
-        num_new_instances += _VSDetectRegistry.detect()
-
-        if num_new_instances > 0 or cls.msvs_manager is None:
-
-            msvc_instances = []
-            msvc_instances.extend(_VSDetectVSWhere.msvc_instances)
-            msvc_instances.extend(_VSDetectRegistry.msvc_instances)
-
-            msvs_instances = []
-            msvs_instances.extend(_VSDetectVSWhere.msvs_instances)
-            msvs_instances.extend(_VSDetectRegistry.msvs_instances)
-
-            msvc_installed = MSVCInstalled.factory(
-                msvc_instances=msvc_instances,
-            )
-
-            msvs_installed = MSVSInstalled.factory(
-                msvs_instances=msvs_instances,
-            )
-
-            cls.msvs_manager = MSVSManager.factory(
-                msvc_installed=msvc_installed,
-                msvs_installed=msvs_installed,
-            )
-
-        return cls.msvs_manager
-
-def vs_detect(env=None):
+def msvs_manager(env=None):
     vswhere_env = env.subst('$VSWHERE') if env and 'VSWHERE' in env else None
-    msvs_manager = _VSDetect.detect(vswhere_env)
-    return msvs_manager
+    vs_manager = MSVC.VSDetect.msvs_manager(vswhere_env)
+    return vs_manager
 
 def _find_msvc_instance(msvc_version, env=None):
 
@@ -3408,7 +770,7 @@ def _find_msvc_instance(msvc_version, env=None):
         debug("Unknown version of MSVC: %s", repr(msvc_version))
         return msvc_instance, query_key
 
-    vs_manager = vs_detect(env)
+    vs_manager = msvs_manager(env)
 
     msvc_instances, query_key = vs_manager.query_msvc_instances(msvc_version=msvc_version)
 
@@ -3536,7 +898,7 @@ def _reset_installed_vcs() -> None:
     debug('reset')
 
 # register vcs cache reset function with vs detection
-_VSDetect.register_reset_func(_reset_installed_vcs)
+MSVC.VSDetect.register_reset_func(_reset_installed_vcs)
 
 _VC_TOOLS_VERSION_FILE_PATH = ['Auxiliary', 'Build', 'Microsoft.VCToolsVersion.default.txt']
 _VC_TOOLS_VERSION_FILE = os.sep.join(_VC_TOOLS_VERSION_FILE_PATH)
@@ -3713,27 +1075,19 @@ def _msvc_instance_check_files_exist(msvc_instance) -> bool:
 
     return False
 
+# TODO(JCB): register callback (temporary until toolset detection is added)
+MSVC.VSDetect.register_msvc_instance_check_files_exist(_msvc_instance_check_files_exist)
+
 def get_installed_msvc_instances(env=None):
     global _cache_installed_msvc_instances
 
     # the installed instance cache is cleared if new instances are discovered
-    vs_manager = vs_detect(env)
+    vs_manager = msvs_manager(env)
 
     if _cache_installed_msvc_instances is not None:
         return _cache_installed_msvc_instances
 
-    msvc_installed = vs_manager.msvc_installed
-
-    # TODO(JCB): first-use default channel
-    vs_channel_def = _VSChannel.get_default_channel()
-
-    msvs_channel_map = msvc_installed.msvs_channel_map
-    msvc_version_map = msvs_channel_map.get(vs_channel_def)
-
-    installed_msvc_instances = [
-        msvc_version_map[vc_version][0]
-        for vc_version in msvc_version_map.keys()
-    ]
+    installed_msvc_instances = vs_manager.get_installed_msvc_instances()
 
     _cache_installed_msvc_instances = installed_msvc_instances
     # debug("installed_msvc_instances=%s", _cache_installed_msvc_versions)
@@ -3743,23 +1097,12 @@ def get_installed_vcs(env=None):
     global _cache_installed_msvc_versions
 
     # the installed version cache is cleared if new instances are discovered
-    vs_manager = vs_detect(env)
+    vs_manager = msvs_manager(env)
 
     if _cache_installed_msvc_versions is not None:
         return _cache_installed_msvc_versions
 
-    msvc_installed = vs_manager.msvc_installed
-
-    # TODO(JCB): first-use default channel
-    vs_channel_def = _VSChannel.get_default_channel()
-
-    msvs_channel_map = msvc_installed.msvs_channel_map
-    msvc_version_map = msvs_channel_map.get(vs_channel_def)
-
-    installed_msvc_versions = [
-        vc_version
-        for vc_version in msvc_version_map.keys()
-    ]
+    installed_msvc_versions = vs_manager.get_installed_msvc_versions()
 
     _cache_installed_msvc_versions = installed_msvc_versions
     debug("installed_msvc_versions=%s", _cache_installed_msvc_versions)
@@ -3768,10 +1111,6 @@ def get_installed_vcs(env=None):
 def reset_installed_vcs() -> None:
     """Make it try again to find VC.  This is just for the tests."""
     _reset_installed_vcs()
-    _Util.reset()
-    _VSWhere.reset()
-    _VSChannel.reset()
-    _VSDetect.reset()
     MSVC._reset()
 
 def msvc_default_version(env=None):
@@ -4151,16 +1490,15 @@ def msvc_setup_env(env) -> None:
 
     # final check to issue a warning if the requested compiler is not present
     if not found_cl_path:
-        if CONFIG_CACHE:
-            propose = f"SCONS_CACHE_MSVC_CONFIG caching enabled, remove cache file {CONFIG_CACHE} if out of date."
-        else:
-            propose = "It may need to be installed separately with Visual Studio."
         warn_msg = "Could not find requested MSVC compiler 'cl'."
+        if CONFIG_CACHE:
+            warn_msg += f" SCONS_CACHE_MSVC_CONFIG caching enabled, remove cache file {CONFIG_CACHE} if out of date."
+        else:
+            warn_msg += " It may need to be installed separately with Visual Studio."
         if found_cl_envpath:
             warn_msg += " A 'cl' was found on the scons ENV path which may be erroneous."
-        warn_msg += " %s"
-        debug(warn_msg, propose)
-        SCons.Warnings.warn(SCons.Warnings.VisualCMissingWarning, warn_msg % propose)
+        debug(warn_msg)
+        SCons.Warnings.warn(SCons.Warnings.VisualCMissingWarning, warn_msg)
 
 def msvc_exists(env=None, version=None) -> bool:
     vcs = get_installed_vcs(env)
@@ -4434,25 +1772,6 @@ def msvc_query_version_toolset(version=None, prefer_newest: bool=True):
     msg = f'MSVC toolset version {version!r} not found'
     raise MSVCToolsetVersionNotFound(msg)
 
-def _verify() -> None:
-
-    def _compare_product_sets(set_config, set_local, label) -> None:
-        diff = set_config - set_local
-        if diff:
-            keys = ', '.join([repr(s) for s in sorted(diff, reverse=True)])
-            errmsg = f'{label} missing keys: {keys}'
-            debug('MSVCInternalError: %s', errmsg)
-            raise MSVCInternalError(errmsg)
-
-    vswhere_config = MSVC.Config.VSWHERE_SUPPORTED_PRODUCTS
-    vswhere_local = set(_VSConfig.DETECT_VSWHERE.keys())
-    _compare_product_sets(vswhere_config, vswhere_local, '_VSConfig.DETECT_VSWHERE')
-
-    registry_config = MSVC.Config.REGISTRY_SUPPORTED_PRODUCTS
-    registry_local = set(_VSConfig.DETECT_REGISTRY.keys())
-    _compare_product_sets(registry_config, registry_local, '_VSConfig.DETECT_REGISTRY')
-
 # internal consistency checks (should be last)
 MSVC._verify()
-_verify()
