+++ test/fixture/no_msvc/no_msvcs_sconstruct_msvc_toolset_versions.py
@@ -11,9 +11,9 @@ def DummyVsWhereExecutables(vswhere_env=None):
     # not testing versions with vswhere, so return empty list
     return []
 
-for detect_cfg in SCons.Tool.MSCommon.vc._VSDetectRegistry.DETECT_CONFIG.values():
+for detect_cfg in SCons.Tool.MSCommon.MSVC.VSDetect._VSDetectRegistry.DETECT_CONFIG.values():
     detect_cfg.vc_cfg.regkeys.clear()
 
-SCons.Tool.MSCommon.vc._VSWhere.find_executables = DummyVsWhereExecutables
+SCons.Tool.MSCommon.MSVC.VSWhere.vswhere_get_executables = DummyVsWhereExecutables
 toolset_version_list = SCons.Tool.MSCommon.msvc_toolset_versions()
 print('toolset_version_list=' + repr(toolset_version_list))
