+++ test/fixture/no_msvc/no_regs_sconstruct.py
@@ -7,7 +7,7 @@ import SCons.Tool.MSCommon
 
 DefaultEnvironment(tools=[])
 
-for detect_cfg in SCons.Tool.MSCommon.vc._VSDetectRegistry.DETECT_CONFIG.values():
+for detect_cfg in SCons.Tool.MSCommon.MSVC.VSDetect._VSDetectRegistry.DETECT_CONFIG.values():
     detect_cfg.vc_cfg.regkeys.clear()
 
 env = SCons.Environment.Environment()
